This commit may need to be reversed, for some of the older compilers but I'm guessing its safe.
Basically it fixes a warning about typedef usage.
Reading the docs on gcc-3.2 it says that classes that use a typedef
need to have the keyword "typename" infront of them according to the c++
standard.  Currently it only produces a warning but it may be an
error later, so I figured why not fix it.

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@359 954f8c5b-7b00-dc11-b283-0030488c597c