Fix for bug #70.
I am a little skeptical about this patch, because somehow
the file had the TXT_ISEXT flag set in it, and the support for
that feature isn't complete. What I want to know is how that
flag got set in the first place, none of the interface code
turns it on.

At least it fixes the crash.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@566 954f8c5b-7b00-dc11-b283-0030488c597c