* fixed bug "turntable deviation" #364
In fact, I made it working different, much more cooler!

-------The BallTable Method TM (patent pending)------------

This is a hybrid of the old turntable method (a virtual camara that rotates
around a table) and the trackball method. Mouse moving horizontal, gives
the turntable effect, but with vertical movement it works like a trackball.
So, vertically moving in the middle of the window exactly mimics old
turntable method, moving vertically in the window edge gives a 'twist'
rotation.

Try the turntable for fun! (is in user settings menu)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@1018 954f8c5b-7b00-dc11-b283-0030488c597c