Trying again to fix build problems with traditional makefiles.
I took a look at how other BPY_* functions were working in blenkernel/
and got to bad level calls (ah, so this is what that is for...).
As a test, I added BPY_clear_script to the "stubs", argh.  If this works,
I'm curious: are these bad level calls needed only by some targets (irix)
because of peculiarities in compiler linkage?

Thanks again, Chris, if this doesn't work, I'll move or copy  the function to
script.c, probably, also adding Python.h to it.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@1730 954f8c5b-7b00-dc11-b283-0030488c597c