BPython: - Fixed small typo in doc file Object.py, reported by user Bill. - Fixed bug #803:
Bug #803's Summary: print, type() or BezTriple broken

Trying to print type(BezTriple) was crashing Blender because
that type wasn't being initialized to PyType_Type. The other
problem is that the getName method for ipocurves only works for
object ipos.  A warning was added for when users try to use it
with other blocktypes,  we'll later update ipo related stuff in
bpython to work for all types.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@1800 954f8c5b-7b00-dc11-b283-0030488c597c