Script menus: -- trying a different approach to see if the crash on Windows goes away.
Running a script with PyRun_File -- a Python/C API function -- was probably
the cause for crashes on Windows, because it uses a pointer to a FILE
struct and on windows this struct can be "different and incompatible"
depending on which libc was used to build the program.  This is mentioned in
the Python/C API Ref Manual, chapter 2.

Now we're loading the file contents to a string buffer and using PyRun_String.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@1955 954f8c5b-7b00-dc11-b283-0030488c597c