Added a non-proportional mode to loopcut:
Short usage info:
In loopcut mode after selecting a 'reference edge' and you are selecting the % to cut press p to toggle non-proportional mode.

the base side of the reference edge will be indicated by a large pink vertex. This is the side that 0% is calculated from and the side that the non-proportional  cut will mimic. To change the side of the base, press the Fkey

The header has always shown the % of the cut, but now in non-proportional mode, it will show the distance from the base side of the loop. It will also show which cut mode you are currently in.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@2145 954f8c5b-7b00-dc11-b283-0030488c597c