The Anim and Tiles settings in the Image Window are now available again, in the new Properties panel (was first called Transform Properties). They include the same settings as were previously found in the window header: Anim toggle, Start/End frame, Speed, Tiles toggle, X repeat and Y repeat.
The Tiles button used to work incorrectly in 2.25, this is now fixed too.

The drawing of the UV Vertex transform buttons in the Properties panel used to
give popup error messages when in Edit Mode (uv_tface_editing_is_allowed is not
silent). Also these buttons would still draw in Edit Mode. Both are problems
are solved now.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@2386 954f8c5b-7b00-dc11-b283-0030488c597c