Recoded the methods how normals get corrected and flipped during render conversion. Old method tried to solve it per object, detecting a flipped matrix. Since we have displacemapping and raytracing this is too inaccurate. Also, flipping normals should alsways result i correct flipped vertex- normals too.
Solution is to completely recalcuate the 'puno' (vertex normal) flags in
renderHelp.c now (set_normal_flags). The face normals themselves are not
flipped until that stage.

Second major improvement involves the calculation of correct splitted
triangles, when a quad is not flat. It now creates triangles based on the
vertex normals, which have actual information on surrounding geometry, thus
resulting in much smoother results.
Without this you get errors in ray-shadow, with light shining almost
parallel to a face. But again, displacement mapping improves with it too.

Third fix is that by removing normal flipping from displace code the result
is much better now.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@2600 954f8c5b-7b00-dc11-b283-0030488c597c