* the installer script so you can do 'scons wininst' (on win32 only, of course)
This installer script is modified from the original one so that the SConstruct
will dynamically add the files that need to go into the installer. This means
that adding a script to release/scripts will automatically put it in the installer.

Also, adding new resource files in bin/.blender/ will make sure that those get
added to the installer.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@2799 954f8c5b-7b00-dc11-b283-0030488c597c