Fixing a blenderplayer compile error reported by Meino Christian Cramer: - added mainqenter to stubs.c so it can be used in blender/blenkernel/intern/blender.c's setup_app_data.
We can't run the onload scriptlink in setup_app_data because the visible areas won't have been updated and redrawn for the loaded .blend file until control returns to screenmain() in blender/src/editscreen.c.  So an ONLOAD_SCRIPT event is entered in setup_app_data and caught in screenmain, where the onload script is then executed.  All for a good looking demo mode ...

BPython:
- Added image.getBindCode() method, contributed by Andrew Corrigan (thanks!).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@2823 954f8c5b-7b00-dc11-b283-0030488c597c