This fix, bug #1342, makes blender consistant, but won't really solve weakness in design.
Issue was: if you delete the 'mother metaball', which is an object with
name without a number, the entire metaball family isn't polygonized. Sofar
so good, where it not that rendering still worked, because it searches all
objects in the scene for a family member with lowest name.

Doing such a search within the main drawing loops will become a major
slowdown, so this better isn't done.

I don't mind sticking to (documented) consistant behaviour. Alternative is
a flag in object to denote it's a Mother Ball, and check for that instead.
This isn't nice to code in current Blender though... we don't have a signaling
system where to insert such tests reliably, making it very hackish.

Jiri; I hope you can agree with the solution! :) Just want to have tracker
cleaned up, refine it to real bugs.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3251 954f8c5b-7b00-dc11-b283-0030488c597c