Bug fix #1678
Deformed meshes (by armature for example) didn't get drawn correctly
for vertex paint, weightpaint, UV faceselect. Was OK in 2.34 :)

Additional to that I found out vertex paint doesn't correctly use the
'Area' option, when a mesh was deformed. That was an oldie.

Plus; made tooltips for 'Area' and 'Normals' in Paint Panel clear.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3376 954f8c5b-7b00-dc11-b283-0030488c597c