BPython: - Small doc update in a script; - Fixed bug #1742: http://projects.blender.org/tracker/?func=detail&atid=125&aid=1742&group_id=9
It was an internal error in bpython.  I was using G.main->script.last to find the currently running (if any) script, but this isn't reliable, we must check each open script to find out if one of them has the SCRIPT_RUNNING bitflag set.

Thanks intrr for reporting and blendix for pointing how to reproduce the bug.  From my tests it should be working fine now.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3425 954f8c5b-7b00-dc11-b283-0030488c597c