Bug #1909
When choosing "render engine" in Scene Buttons, the newly added or removed
Panels didn't invoke a re-alignment event yet.
Also added code that inserts new panels as good as possible on their
previous locations. This works reliable for 1 new panel, not for more, this
because a Panel only stores its old location, not the locations of all
Panels in a given configuration. Consider that minor issue...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3557 954f8c5b-7b00-dc11-b283-0030488c597c