Bug #1970
Objects being vertex-parented to a mesh, which in turn was being deformed
by a curve or lattice, were never update correctly.
Whilst this is a typical issue for our beloved new Dependency Graph, I
consider it worth at least solving for now. An issue since Blender 1.0
you know. :)

Solution is only working for this specific case; vertex-parent to deformed
object.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3606 954f8c5b-7b00-dc11-b283-0030488c597c