Bugfix #1262
Raw fix for the drive letter issue with Quicktime/Windows filenames.
When a drive letter is missing in a filename, it'll use the drive where the executable is.

Fix might be useful for Yafray too, because the extern char bprogname[]; doesn't return
a full path when Blender is started from the console.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3628 954f8c5b-7b00-dc11-b283-0030488c597c