Bug fix as reported by Fish-Face/Chris Le Sueur on IRC
After deleting an object, the call free_and_unlink_base() takes care of
NULLifying object pointers. But, with exception of constraints. Silly!!!

So, in the end of a function additionally the test_scene_constraints()
has to be called to prevent crashes. This was missing in join_mesh and
convert_object.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3935 954f8c5b-7b00-dc11-b283-0030488c597c