*sigh*  I obviously rushed this.  I honestly thought that this commit would not affect anyone.
Changed:
if user_openexr == 'false':
	defines += ['WITH_OPENEXR']

To what it was suppose to be:
user_openexr = 'false':
if user_openexr == 'true':
	defines += ['WITH_OPENEXR']

This was messed up on most platforms, somewhere when adding them all
I switched to the bad version and continued with that...

I'm on irc if anyone else has any commits to clean this up.  I'm going
through things and trying to make sure its working for everyone.

Sorry,

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3938 954f8c5b-7b00-dc11-b283-0030488c597c