NumInput is now applied correctly to constraints and prints correctly in the header.
What that means is that you can grab/resize an object, press Xkey, type 2 (to resize/move by 2 on the x axis) and then press Ykey and the numeric value will be applied to the transformation on the Y axis.

Fixed a small glitch where constraint center wasn't readjusted for edit mode when using MMB (graphical bug only).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3939 954f8c5b-7b00-dc11-b283-0030488c597c