New AFFECTALL flag for numinput When that flag is set, numbers typed when the cursor is on the first position will affect all positions who do not currently have something typed for them.
Enabled Skey, 2, Enter to quickly double the size of something.

All modesty asside, it's coded rather smartly so it shows the cursor in every position that it will affect and lets you tab in undefined position to type in values there.

Thanks to Samadam for reminding me that the old code permitted that.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3946 954f8c5b-7b00-dc11-b283-0030488c597c