xcode build system for blender :
only blender itself for the moment, no player.
international build

divided in 3 subs :

- makesdna
- blender
- bundle packaging

build in order.
you should be able to enable CVS from inside xcode, even for the project,
but you need an initial checkout.

Note this is an initial project, which should evolve to something more
configurable. However, the main and officials build systems stay as
make and scons. This is only an help for the users.

Apple fix&continue feature cannot work due to the way we build the app


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@3952 954f8c5b-7b00-dc11-b283-0030488c597c