New feature, since long on todo;
The grid function (transform translate, snap-to) now uses the grid step as
displayed in the 3d window, so it depends on how much you zoomed in/out.

The threshold for it is a bit tricky... it follows the drawing, but the
blending in/out of grid makes it sometimes not obvious.  might need slight
tweak?


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4024 954f8c5b-7b00-dc11-b283-0030488c597c