TEMPORAL HACK!!!
Added the is_a_really_crappy_nvidia_card() call in BMF_DrawString(), this
to solve a bug in NVidia 6800 drivers of MacOSX G5. It is #ifdeffed for
OSX only, and queries for a NVidia 6800 card to activate the patch.

The issue is that these drivers forgot to correctly implement viewport()
offset for drawing bitmap fonts, causing text display in Blender to be
invisible, except for the leftmost/bottom sub window.

This hack will be removed when Apple releases a driver upgrade, which is
unknown when to happen. Has to be decided still if this is worth for a
release, or that we provide the hack as separate download.

Thanks Randall Rickert for all testing, and Daniel for code review! :)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4056 954f8c5b-7b00-dc11-b283-0030488c597c