Bug fix #2535
Armatures with > 100 bones can deliver a 'parent IK menu' that doesn't
fit on screen anymore. Fixed it with changes;

- when menu doesnt fit to right/left, it places it aligned to left hand
  of window
- when more than 8 collums in a menu exist, it makes collums 25 items
  instead of 20.

Also fixed for the popup menu on Bone parenting. On a 1280 screen we now
can have a menu with over 300 bones (even more when you zoom down the
buttons window a bit). Still not optimal, but at least it's less annoying.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4422 954f8c5b-7b00-dc11-b283-0030488c597c