Transform MMB code fixing Ton's last fix.
His fix was good, except that it switched to project_float instead of project_short, and that's not good because we're renormalizing the vector after that, which maximizes the float errors. Since it's scalling the axis anyway with the zfac method, it can use project_short safely (I tested it with the bug tracker file that triggered the fix in the first place).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4436 954f8c5b-7b00-dc11-b283-0030488c597c