Changes to Documentation Format ONLY. no executable code.
Martin noticed many of our bpy instance variables were
incorrectly marked as class variables in the doc.  This
commit essentially changes the title of sections of the doc
from Class Variables to Instance Variables.  Now that we are
adding class or module dictionaries for constants, etc. this is
a distinction worth making.  Plus it is right.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4605 954f8c5b-7b00-dc11-b283-0030488c597c