Work for fix of #2256 "something weird with some buttons":
NUMSLI buttons with retval of 0 can be used to trigger other buttons. Not
sure if this is best course of action but will investigate more. Until then
buttons_shading.c is clear of NUMSLI buttons with retval of 0 and now use
B_DIFF as retval argument (which happens to be 1).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@4988 954f8c5b-7b00-dc11-b283-0030488c597c