First version of visualizing the DoF (french degrees!) for Pose-bones. It now only draws the limits for X and Z rotations (Y is bone axis itself).
Quick snaphsots (links will disappear):
http://www.blender.org/bf/rt6.jpg
http://www.blender.org/bf/rt7.jpg

It only draws for selected Bones that are part of IK, and have limits set.
Most work was getting code OK to setup drawing this 'DoF space', so now
experiments can be done with more drawing types.

ALso; Buttons for DoFs now only draw if the pose-bones are part of an IK
chain.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5222 954f8c5b-7b00-dc11-b283-0030488c597c