Memory management flag to behave better, as hinted by Ken Hughes. Still not good, i.e. getting these when quitting: Error Totblock: 4 new bpytriple len: 60 0x8889bdc ... 'cause nothing frees them..
Changed the loop that parsed input args to PyArg_ParseTuple to have
support for passing ints from Python too as the floats that are the
coordinates. Didn't find PyInt_AsFloat and figured that this is an ok
way anyhow.

Changed the default handle mode from AUTO to ALIGN, which is the same
as in UI and more useful at least for me.

Little sanifying in CurNurb (this was done with Ton).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5279 954f8c5b-7b00-dc11-b283-0030488c597c