Bugfix #3077
Radiosity didn't add edges block, on "Replace Meshes". Caused wireframe
not to draw, but also crashes like for join().

Also: added patch that sets the drawflags in edges derived from the
ones set in faces. This ensures the conversion to be done correct.

TODO: this edges call also used in other areas in code, that has to be
catched and done differently.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5320 954f8c5b-7b00-dc11-b283-0030488c597c