Bugfix #3085
When you move the mouse, while dragging numbers in buttons (or sliding),
and release it over another button, the release event was passed on to that
button. It happened with Radiosity buttons for example.

Reason; buttons with return event 0 are passing on events to other buttons.
This is needed for LABEL or ROUNDBOX buttons, to prevent them blocking
events when there's another button inside.

Made sure all these number/slider buttons returning B_NOP now.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5328 954f8c5b-7b00-dc11-b283-0030488c597c