New! Hooks now support vertex groups. With weight painting it looks very interesting. Almost z-painting! :) Try a sphere with all vertices in a group, hook it, and scale empty in object mode a bit. http://www.blender.org/bf/rt3.jpg (Which gives the idea to make a Displace Modifier working with normals and vertexgroups..)
Works as follows now; in Mesh editmode, when no vertices are selected, it
tries to assign on CTRL+H command the active Vertex Group. (Menu is getting
to long now... need to think over). Only works when vertices are assigned
to a Vertex Group, this to calculate the Hook center.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5351 954f8c5b-7b00-dc11-b283-0030488c597c