Removed ancient (2.1) feature for Constraints... a highly undocumented and unpredictable one!
This is what it did; if two or more Constraints are of the same type, it
accumulates the Target positions/rotations for the constraints, averages
them, and then only applies the last Constraint in the row. It seems to
be a trick to blend IK Constraints or so... in all other cases I cannot
find a good use for it (nor did Bassam).

For example; add three Empties, and make one Empty to have 2 location
constraints to the other two. This just didn't work, unless you insert
an "Empty" constraint inbetween.
I will post in the blender.org animation forum feedback for it too. :)

Its quite easy to make it an option, but I first like to grasp fully what
the actual use of such an option is.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5376 954f8c5b-7b00-dc11-b283-0030488c597c