New: Hotkey/menu access in 3D window to add constraints. Works in PoseMode as well as for Objects. Hotkey: CTRL+ALT+C (bit clumsy I know, but I like to use the Ckey).
Constraints are added to the active Object or Bone.
Based on selection context, the menu has three versions:

Pose Mode:
  - if another bone is selected, bone becomes target
  - else if another Object is selected, Object becomes target
  - else it adds a new Empty as target
Object Mode
  - if another Object is selected: Object becomes target
  - else it adds a new Empty as target

Same works this way now for CTRL+I, "Add IK".

To be solved: ALT+C in PoseMode is "Clear Constraints", but in Object Mode
it does "Convert" still...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5548 954f8c5b-7b00-dc11-b283-0030488c597c