Bugfix #3304
Actually nice feature too: when using Scene "Set" option, that scene draws
its transparent and x-ray stuff independent of the 'foreground' scene.
That means that x-ray in a set will be always over the set objects, but
stays behind the current scene.

For transparent it works the same though... meaning that solid objects from
the foreground scene will always draw over transparent from the "Set".

Note; this only is true for the 3d window, or the OpenGL render, not for
actual rendering of course.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5693 954f8c5b-7b00-dc11-b283-0030488c597c