*armature api for python - don't get too excited - allows you to get armatures from a scene - makeEditable()/saveChanges() puts the armature into out of editmode (pythonically) - Armature.bones is a dictionary that contains all the bones in the armature and can be iterated - getters are available for: name, roll (dictionary) keys are BONESPACE, ARMATURESPACE head (dictionary) keys are BONESPACE, ARMATURESPACE tail (dictionary) keys are BONESPACE, ARMATURESPACE matrix (dictionary) keys are BONESPACE, ARMATURESPACE weight deform_dist subdivisions options (list of constants) parent children
Setter work only in editmode. Some are not fully implemented.
Type class is embedded in the module. This means the construct is called as follows:
Blender.Armature.ArmatureType()

import Blender.Armature as Armature
arm = Armature.Get('myarm')
for name, bone in arm.bones.items():
...print name, bone, bone.matrix['ARMATURESPACE']

more documentation is forth coming. This is an alpha for this api.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5727 954f8c5b-7b00-dc11-b283-0030488c597c