Patch provided by Stephan Kassemeyer.
This fixes an error in the striding system, which was by default correcting
path position backwards in time, which could give errors when a stride bone
moves in the beginning of a path faster than the path moves. (Can you
follow that? cool!)

The patch checks for this case, it corrects by default forwards in time,
unless we're at the end of a path. As bonus this fix will also ensure the
character stays on the path closer.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@5846 954f8c5b-7b00-dc11-b283-0030488c597c