Bugfix #3580
Apparently python allows to create Key blocks without proper *from pointer
back to the owner of the Key. The Key unfortunately wasn't mean to be a
dynamic block linkable to any Mesh...

This patch sets the from pointer on any call to do_mesh_key, which is safe
to do anyway. I made a note to get rid of that convention once, but at
least now we got non-crashing blenders. :)

(In bugreport noted that we need to check Python code for it too!)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@6100 954f8c5b-7b00-dc11-b283-0030488c597c