after much suffering, got GE build and work almost cleanly on Os X (with make, need to confirm with scons)
after cleaning the changes are in fact minimal, but the situation
is still quite a bit hackish.

Game engine coders, there is also quite a number of warnings that
need to be fixed.

current situation is that everything seems to work, but GLSL shaders
spew a lot of errors on console and blender may crash on exit when
a GLSL shader was used. ARB stuff works fine.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@6466 954f8c5b-7b00-dc11-b283-0030488c597c