Fix for bug #3886
Basically the check for a type of jpeg was messed up.
I added a imb_is_a_jpeg function since I was working on that bit and
it makes it a little more similar to the other filetypes.

I also changed a switch statement that had the same loop for all
cases, just moved the loop outside of the switch statement.

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@6945 954f8c5b-7b00-dc11-b283-0030488c597c