fix for bug #3930 Colour picker broken - HSV sliders
Note : this part of code was quite a mess with things called twice when
not needed, tests for not updating when we want to, and hsv sliders not
handled properly.

I tried to figure what was useful and what was not but cannot garantee i
spotted everything or even fixed it the most proper way, the event callback
nesting making hard to figure what really happen.

However it does work again now


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7050 954f8c5b-7b00-dc11-b283-0030488c597c