Bugfix #3961
When using CTRL+L for materials, you can end up with non-existing material
indices in faces. The drawing code then was still happily drawing the old
situation (or something random, its a static array).

This commit checks the maximum amount of materials on an object, and draws
the last available material for a non-existing index. It uses an ugly
global yes, but this code is bad anyway. :)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7227 954f8c5b-7b00-dc11-b283-0030488c597c