Small update for pydrivers: force reloading the pydrivers.py Blender text module when user edits the input text box of any pydriver (Transform Properties panel, Ipo window).
It's enough to click in and out of a single pydriver's text input box
for the module reloading and also re-evaluation of all pydrivers
available. Maybe this "refreshing" should also be available from a
menu, let's see.

Note for Python fans:

Definitions and redefinitions in a reloaded module are properly handled
in Python, but previously defined data in the module doesn't disappear.
So if you define a function "f" inside a module, import it, then change
the function's name to "g" and reload the module, both "f" and "g" will
be available. This is considered a feature, check reload's documentation:
http://docs.python.org/lib/built-in-funcs.html#l2h-59


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7339 954f8c5b-7b00-dc11-b283-0030488c597c