Bugfix #4298
Uncommitted the feature I added for Plumiferos to be able to render
animated render-borders (which effectively means every frame can be a
different size).
For several reasons;

- it will crash movie rendering (when image sizes differ)
- it forced code to do a full initialize each frame, also for setting the
  renderwindow (which popped up on each frame)
- the render pipeline was not designed with per-frame python changes in
  mind for image sizes... it uses an "Initialize" stage which is only
  called once for an entire sequence.

That latter might be an omission, but for that I better code a new API
call (for use in Python) so a re-init can be enforced.

The old functionality (animated render borders) will still work when you
render in background, using small steps like blender -s 1 -e 5 -a etc.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7656 954f8c5b-7b00-dc11-b283-0030488c597c