==Mesh modeling==
A modification of how 'Extrude Region' behaves on the line of
symmetry when a mirror modifier is used. If 'Do clipping' is
selected, and there is an edge selected on the line of symmetry
that is connected to a selected face, that edge no longer gets
extruded into a face lying on the symmetry line. In pictures,
here is how the old behavior compares to the new behavior:

http://bebop.cns.ualberta.ca/~cwant/mirror_extrude_region

I think this new behavior is consistent with what most users of
the mirror modifiers would want, but if not please say so!

Also: Test, test, test!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7668 954f8c5b-7b00-dc11-b283-0030488c597c