Bugfix #4363
In windows, without temp path set, the 'save buffers' render option crashes.

I've coded a blenlib BLI_is_writable(char *filename) to check for such
cases. This is not much needed in Blender, since the open() command is
checked for. However, file saving happens deep inside the C++ exr lib, and
it throws an exception crash when a file cannot be written.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7792 954f8c5b-7b00-dc11-b283-0030488c597c