Plumiferos bug: autohandles in Ipos could get screwed up in extreme cases, but they got this case! :)
Bug was that for autohandle, the previous position of handle was used for
calculating the position too, which I really don't remember why... (that
is code from 10 years ago). Problem with that approach is that extreme
changes in handles don't go correct immediately, but need to itterate a
while. Or even worse, can give NaN values, resulting in this:

http://www.blender.org/bf/ipobug.jpg

Now the handle points are fully recalculated, as it should.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7832 954f8c5b-7b00-dc11-b283-0030488c597c