Hooray! Finally found the dreaded "Opengl crash" the poor orange team suffered for the entire movie. :)
It only happened when rendering large frames, using a lot of memory and
typically when you also use other software in meantime.

Reason: the main thread does the drawing updating, while rendering is
still continuing. When using Ztransp, there was a free buffer done
when possibly a draw could still be in progress. Only crashed when drawing
is slow... explaining why it only showed up in more complex cases.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@7984 954f8c5b-7b00-dc11-b283-0030488c597c