Bug #4568
No fix... the Object "TimeOffset" option to work on own ipo doesn't work
at all, not since NaN days. Enabling it is easy, but might screw up a lot
of previous saved files.

I moved that report to the todo, for a make-over of the entire internal
timing system. Added tooltip in button this option doesn't work.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8010 954f8c5b-7b00-dc11-b283-0030488c597c