Bugfix #4556
When a curve guide animates, it doesn't correctly work on dynamic particles.
This because the guides work on a fixed starting position (the birth of
particle) and then defines a full path for the entire particle. It is clear
this was only coded with static particles in mind...

Nevertheless, I've added a line of code to at least put the curve guide
on correct location for the moment a particle gets born.

Report moved to the todo tracker; this should be a spec for particle recode


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8012 954f8c5b-7b00-dc11-b283-0030488c597c