Bug as reported on irc often: in default .B.blend a first opening of Outliner happened in Oops block view. I found this was caused by commit of of a new .B.blend in may 2005 (after 2.36, before 2.37). That file had already opened Oops views stored, which have to remain that way of course.
A full cleanup of .B.blend i rather not do now, so I've added a patch in
code that on read of the built-in .B.blend it sets all existing Oops
windows to Outliner.

And while working on it anyway, on first opening of a new Outliner, the
current Scene level is opened, but showing objects closed.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8050 954f8c5b-7b00-dc11-b283-0030488c597c