Bug fix #4642
When a Sequence setup had a gap, it returned default black frames. However,
this black frame was overriding the float buffer when a new strip started
to render again, so the Sequencer kept saving black.

Also: added extra info print for Scene strip, this didn't tell yet what the
start/end frame of Scene was, and what current frame was.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8120 954f8c5b-7b00-dc11-b283-0030488c597c