Bug fix #4627
The transform number input code allowed to keep typing forever. That's nice
but the code uses floats to store values, so there's a limit of 7 digits
resolution to take care of. I've added this limit now, it will stop when
the 8th digit was typed, giving a range of 100 million, quite OK for
Blender measures.

There was also a short in use for values between 0 and 1, causing an error
when you try to type like 0.99999. Here I've added a limit of 7 digits.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8122 954f8c5b-7b00-dc11-b283-0030488c597c