Bugfix #4770
When adding a new node in a group, the call to refresh input/output
sockets was called after a compositing update was executed.
Just moved this call 2 lines up.

This also uncommits the fix from Joilnen, the stack pointers in the node
system are *per definition* set. If NULL, it's an indication something
else is wrong.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8203 954f8c5b-7b00-dc11-b283-0030488c597c