--Value Squeeze Node--
This material node takes an input value of any size and fits it to a sigmoid curve (value between 0 and 1) The width of the curve can set to make the sigmoid fall off faster (bigger number) or slower (smaller number) and the centerpoint (what value is mapped to 0.5) can be adjusted as well. Anyone smarter than me can feel free to tweak this and make it better.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8272 954f8c5b-7b00-dc11-b283-0030488c597c