Bugfix #4923
Particle distribution error when using Vertex Weights.
In the unlucky case the last face in a Mesh has no weight (zero), the code
that tries to distribute particles evenly accidentally could assign it
still a particle, because of the 'remainder' calculus.

Also only happens in cases where the remainder has a value.
So a workaround is also to increase/decrease amount of particles.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8387 954f8c5b-7b00-dc11-b283-0030488c597c