== 3D View ==
Customizable Grid Subdivisions

This commit adds a numbut to the View Properties panel that lets you specify how the grid is subdivided.

This affects snapping in translations, obviously.

Default: 10 (behavior doesn't change)

That means people still stuck in feet and inches (shudders) can set it to 12 and have 1 unit = 1 foot.

That also means you can work in "heads" when doing body proportions or whatnot (don't think of it as being limite to "CAD" uses).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8677 954f8c5b-7b00-dc11-b283-0030488c597c