More Softbodies
Added a less 'strict' solver step size control.
Cuts down needed loops to a 1/5 in 'regular' collision situations.
So why be a maths smartass when we want to do a skirt in a breeze :)
Is on by default, old mode can be turned on with the 'O' button beside the RK-limit.

Some minor fixes to the UI


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8909 954f8c5b-7b00-dc11-b283-0030488c597c