Yuck, a very ancient crasher in going area-full screen in Blender. I wonder how this survived so long in Blender... (2005/03/09 commit).
Reason was the call to BPY_free_screen_spacehandlers(sc) in kernel,
which was freeing up scripthandlers in a weird way. That call is
really obsolete. The real freeing should go in the del_area() call,
to prevent copying and deleting area in the UI to go wrong.

(Crash happened in testing timeline markers, and holding CTRL+Uparrow
a while...)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@8926 954f8c5b-7b00-dc11-b283-0030488c597c