=== Interface ===
Based on patch #5140 by Juho Vepsäläinen, this commit removes the requirement to type # at the start when you want to use Python expression evaluation when typing a value in a button.

In a nutshell, that means you can now type 3 + 5 in a numbut and see it change to 8.

Word of warning: The normal Python operator logic applies, so if you type in 1 / 3, you'll get 0 and not 0.333. There's no going around that.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9162 954f8c5b-7b00-dc11-b283-0030488c597c