Bugfix #5434: Animation (alt-a / timeline) will not play if Sta: frame set higher than End: frame
The end frame should not be allowed to have a value less
than the start frame. This commit sets the minimum allowable
value of the end frame button in the Anim panel and the timeline
to the start frame value.

Ton/Matt - if there is a good reason to not do this, feel free to
revert it back.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9324 954f8c5b-7b00-dc11-b283-0030488c597c