== Action Editor - 'Long' Keyframes ==
When animating, it is often useful to be able to visually see where the
'pauses' are between keyframes. Long keyframes do this - linking two
keyframes in the same channel together.

Long keyframes are only drawn when the two keyframes have the exact
same values. This has to happen for every ipo-curve represented by the
keyframes shown for a long keyframe to be drawn.

I've added two new theme colours for the action editor. They are for
the selected and deselected colours of the long keyframes (currently
defaulted to be the same as the NLA strip selection colours).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9327 954f8c5b-7b00-dc11-b283-0030488c597c