=IDProperties Python Test GUI Script update=
Stupid typo in script, you couldn't go into
a subgroup if it was the first property in 
the list.

Also, name collision code needed a little work.
This could result in overwriting properties
with the New button, which should instead
work as the rest of blender and add .XXX
to the name.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9346 954f8c5b-7b00-dc11-b283-0030488c597c