== Editmode Bone Snapping ==
I've often found it very annoying that with both ends of a bone selected,
they would both get snapped to the snapping point. This means that the bone
becomes zero-length, hence disappears from view and gets deleted upon
leaving editmode.

Now, when both ends of the bone are selected, only the head of the bone gets
snapped to the snapping point. The tail will get offset by the same amount
that the head gets offset by, thus preventing zero-length bones.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9493 954f8c5b-7b00-dc11-b283-0030488c597c