Bugfix #5600:
When clicking on the basis curve of a Curve's shapekeys in the IPO Editor
with 'Relative' turned off, Blender crashed. This was caused by the call to
set_sculpt_object() which only works on meshes (at the moment).

I've added a check to make sure this is only called for meshes. Also, silenced
the 'implicit declaration' warning my compiler was giving me about this function.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9607 954f8c5b-7b00-dc11-b283-0030488c597c