=== Transform Bug Fix ===
[ #5458 ] Some transforms are not accurate while pressing shift key

Shrink/Fatten & Push/Pull: Shift slowed the transformation but didn't add precision. (integer division. fixed by applying the same solution used in Grab)

Warp: Shift was ignored. Fixed by changing warp to use the same input method has To Sphere. That is, the 3D view acts like a big horizontal slider: Left -> No effect, Right -> Full effect. This is somewhat of a "big" change from how warp used to work but this is more predictable and is not limited by the original position of the mouse pointer. (of course, this works with Shift correctly too).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9797 954f8c5b-7b00-dc11-b283-0030488c597c