in seqeffects.c I removed and extra ; and then did a lot of formatting fixes.
gamma.c made it return B_PLUGIN_VERSION instead of a hard coded 4.
I haven't had time to look at the other plugins but I'm guessing
the deal was B_PLUGIN_VERSION was still 3 and gamma.c was float aware
so it was bummped to 4.  Now that B_PLUGIN_VERSION is 5 I'm guessing
the other plugins need to be updated so they are at level 5 as well.
(float aware)  I'm going to work on it later today.  At home with a
sick kid.  This is all in reference to bug #5692

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@9809 954f8c5b-7b00-dc11-b283-0030488c597c