Added headers and a precompiled library for OpenAL on Intel OSX.
There are still *problems* with the functionality that OpenAL needs to
fix for this platform.  Do not expect OpenAL sounds to work properly in
the game engine, for example, and there may be some crasher bugs.

However, this will simplify building and running blender on Intel OSX
if you're not interested in using the OpenAL featuresOpenAL features.
More progress will be made on the makefiles to use this library once it
is less crash-prone.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10195 954f8c5b-7b00-dc11-b283-0030488c597c