=== Export UV Layout ===
[ #6450 ] Save UV Layout: Overflow Error

This error was caused by excentric UV faces being exported (I'm talking things like 3443870976 as UV coord, in a range of 0..1).

Edges with extreme coordinates are now ignored (for TGA only, they are exported to SVG) and a warning is printed (at most once) to the console.

I chose to still export them to SVG because they don't affect the running time of the script while exporting them to TGA is just insanely long.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10379 954f8c5b-7b00-dc11-b283-0030488c597c