== NLA Editor - Auto-Blending for ActionStrips ==
This is just a little time-saver for NLA-workflow. With the 'Auto-Blending' option
turned on in the Transform Properties (NKEY) for an ActionStrip, that strip's blendin/blendout values are determined based on the number of frames that the previous and/or next actionstrip(s) on overlap over the start and end of it.

It is turned on by default for new actionstrips added using the Shift-N hotkey.

Caveats:
* Only the actionstrips immediately on either side of the strip being evaluated, will
 have any effect
* A strip that is longer-than, and extends over the sides of the strip being evaluated,
 will have no effect

Additional Notes:
* Blendin/Blendout have been renamed In/Out in UI for brevity
* Button layout in NLA Transform Properties has changed slightly again, but hopefully that shouldn't be too much of an issue.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10531 954f8c5b-7b00-dc11-b283-0030488c597c