== Preview Range ==
Removing the ability to have the Preview Range affecting the Preview Renders (OpenGL renders) as it is unstable in certain cases. It was only a problem when rendering to movie-file formats with a preview range set greater than the normal range, as it would consume increasing amounts of memory for each frame rendered.

Preventing this from happening would have required too much disruption of the avi
writing code to be worthwhile.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10618 954f8c5b-7b00-dc11-b283-0030488c597c