Armature Drawing Tweak:
I've always found it hard to see the armature bones in 'Line' mode that don't have any constraints/ik/keyframes/etc. colour indications as they were set to the colour of the background. This tweak makes these particular bones draw slightly darker than
the minor-gridlines in the 3d-view, but not as dark as the major gridlines.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10643 954f8c5b-7b00-dc11-b283-0030488c597c