Bugfix #6674:
Action Editor: editing of action-channel keys is impossible if shapekeys are present
for that object.

Also, after toggling the 'swimmer' icon in the IPO editor header,on and then off again, then trying to expand/collapse/select/deselect the 'Shape' action channel, Blender would crash.

Not really a showstopper, but still nice to fix. ;-)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10700 954f8c5b-7b00-dc11-b283-0030488c597c