== Multires ==
Modified the vert_edge_map and vert_face_map arrays in MultiresLevel so that the data is stored only for as long as it is needed -- for adding levels and updating levels. Once the modifications are 
complete, the mapping data is deleted, which provides a large reduction in memory usage for a minor increase in processing time for updating levels.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10906 954f8c5b-7b00-dc11-b283-0030488c597c