== Action Editor - Long Keyframes ==
I've recoded the way long keyframes work a bit, so that more cases are handled accurately. Now, it takes into account the number of IPO-curves that have keyframes in them on the start/end frame of the long-keyframe, instead of just taking the total number of IPO-curves present for the channel being drawn.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@10966 954f8c5b-7b00-dc11-b283-0030488c597c