This is patch #6833 https://projects.blender.org/tracker/index.php?func=detail&aid=6833&group_id=9&atid=127
It adds the option when creating a new "screen" to create an empty one
(one with one window)

While its one more click, I think this is pretty dang useful.  If people
want to revert it feel free, and we can talk about it.
This seems more consistant with the Scene bar though so I figured it was safe 
to add it.

Kent



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@11075 954f8c5b-7b00-dc11-b283-0030488c597c