== Constraints - Important Bugfix ==
I've finally fixed the bug with the Constraint Space Conversion. It was a single matrix multiplication in the wrong order (for local->pose). 

Also, there is more code added for the space conversion process when bones have 'hinge' on. (NOTE: this stuff for hinge bones may still not work really nice yet) 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@11355 954f8c5b-7b00-dc11-b283-0030488c597c