Little code cleanup.
bsystem_time was being called with an extra variable, which was useless. Most of the places that called it, were passing NULL for that variable anyway.

I've also cleaned up that function a bit, but the underlying problems with that part of the code still exist (EVIL GLOBALS that are exported for frame_to_float), for mblur and fields rendering features. That remains for another time.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@11485 954f8c5b-7b00-dc11-b283-0030488c597c