* Change to the working of the left mouse select preference
Previously, the implementation of this was quite awkward, with
the preference swapping mouse buttons globally, rather than just 
selection, as is advertised on the pref. This had the effect of 
changing the painting in sculpt/texture paint/weight paint/etc to 
the left mouse button. This was totally silly, since when using a 
tablet, left mouse select is the sane way to go, but it meant 
that every time you wanted to sculpt or paint, you had to switch 
the mouse buttons around so you could actually use the tablet as 
normal.

This commit fixes that, and makes the preference do just what it 
says, use left mouse for selection (i.e. in object/edit mode) and 
only that.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@11726 954f8c5b-7b00-dc11-b283-0030488c597c