This is patch [#6772] Background image drawing speedup Thanks goto Matthew Plough (meestaplu)
I've tested it on linux with both an nvidia and an ATI card,
I've also tested it on a mac and solaris, as well as through a
remote display.  So I think this one is pretty safe.  
If it causes problems on older systems we can revert it
but I don't think we will have any issues with it.

Kent



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12030 954f8c5b-7b00-dc11-b283-0030488c597c