== Clamp To Constraint ==
Now there's an option for the owner to follow the path of the target cyclically. Previously, if the owner moved past the extents of the side of the bounding-box used for the calculations, the object was placed on the curve at the nearest extent. 

This option is only really useful if the curve itself is cyclic, although you can still use it otherwise. To enable, just turn on the cyclic option. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12149 954f8c5b-7b00-dc11-b283-0030488c597c