* Patch #6877: Zoom to Mouse Position by Fahrezal Effendi (exavolt)
This adds an additional preference to the view and controls section, which uses the mouse's position as the centre of the zoom when zooming in the 3D View with Ctrl-MMB or the mouse wheel. It's very nice for big scenes, thanks Fahrezal!

A nice todo would be to add this for 2D views as well.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12237 954f8c5b-7b00-dc11-b283-0030488c597c