IK ==
Solving is now done independent of scale, by scaling the chain to have a
size of about 1.0. This solves some issues with small or big chains, and
also makes the IK stretch setting independent of scale. The latter breaks
backwards compatibility somewhat, but is an improvement over what it did
before.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12450 954f8c5b-7b00-dc11-b283-0030488c597c