Proper code for bugfix revision 12365 (which was uncommitted today).
Old log Message:
-----------
bug fix, when opening blender with a file (by double clicking or from the 
command line) - the initial undo state would be set to the default scene.
So holding Ctrl+Z would go back to the default .B.blend rather then the 
file that the user opened.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12553 954f8c5b-7b00-dc11-b283-0030488c597c