==Mirror Modifier==
Support for using the axes of a different object as the line of mirror
symmetry for a mirror modifier. As a nice consequence, this allows
"clipping" to arbitrary planes in editmode.

A fun example of using a couple of mirror modifiers and an array
modifier to easily make a nice flower type model is here:

http://bebop.cns.ualberta.ca/~cwant/chocolateC05.blend



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12574 954f8c5b-7b00-dc11-b283-0030488c597c