== Auto-Keyframing - 'Only Needed' Improvements ==
The 'Only Needed' option for Auto-Keyframing now works better with a few of the different transform options. 

Rotating/Scaling using a pivot point that is not the center of the Object/Bone (3d-cursor, active), also inserts location keyframes if the location also changes. If "Only Centers" option is activated, then only location keyframes are inserted.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12639 954f8c5b-7b00-dc11-b283-0030488c597c