== Bone Paths - Auto-Recalc Improvements ==
Added a new function which specially caters for auto-recalculation of paths after updates (such as keyframing). 

Currently, all that differs between this and pose_calculate_path, is that it only recalculates for bones with paths. Also, it only updates what it needs to (minimum frame range affected).

Warning: WIP commit (hopefully it compiles ok)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@12777 954f8c5b-7b00-dc11-b283-0030488c597c