== AutoKeying - Tidy Up ==
This commit restores the "big red button". When Auto-Keying is enabled (record button toggled on), a menu appears beside it to choose which behaviour for auto-keying should be used. This should be more efficient for quickly turning Auto-Keying on/off.

I've also added the relevant version-patches to fix up old files so that sensible default options are in use.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@13215 954f8c5b-7b00-dc11-b283-0030488c597c