== Action Editor - Copy/Paste Improvements ==
Changed the way the offsetting of pasted keys relative to the current time works. Now, all pasted frames are offset by the same amount. This is calculated as the difference in the times of the current frame and the 'first keyframe' (i.e. the earliest one in all channels to be pasted). The 'first keyframe' is found when doing the copying.  

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@13306 954f8c5b-7b00-dc11-b283-0030488c597c