Bugfix #8070: blender crash when assigning driver to bone constraint
When there were no keyframes for a constraint on a bone, trying to add a driver to the influence curve would fail, as a new curve would be added the some mysterious place. The cause was a missing case, as the code assumed that no user would try to add from IPO-Editor in such a way.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@13490 954f8c5b-7b00-dc11-b283-0030488c597c