Further work on the premul option for ton.  This option (which basically tells the renderer and compositor to expect a key image) is now done at the image user level.
This does have some caveats, as image users don't always work
the way I thought they would/should (for example, the same image user
structure is apparently used in the uv image editor for all images,
which is kindof odd).

The UV image editor also now smartly detects if the premul option is
set and draws the image using key alpha, instead of premul

The subversion level was upped to convert the old premul flag, which was at
the image level, to the new one, which is at the image user level.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@13759 954f8c5b-7b00-dc11-b283-0030488c597c