[#8434] Normal manipultor does not align with normals (like in 2.45)
Bring back multivert normal average for the normal manipulator/orientation.

When selecting more than three vertice, the averaged normal of all selected is used to get orientation Z-axis.

NOTE: This is only applicable when NO faces are selected, otherwise it uses the averaged normal of the faces (excluding vertice that aren't part of a face). This is not new behavior

CHANGES FROM 2.45: Selecting 2 or 3 vertice doesn't use normal averaging but uses virtual edge and face orientations instead.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@13962 954f8c5b-7b00-dc11-b283-0030488c597c