-> Color Fixing tool in FLT toolbar
A new tool for for fixing colors on meshes imported from other files with
conflicting color palettes has been added to the FLT toolbar script. This
will find the closest match for all the colors in the faces of selected
meshes and update their baked vertex color representation.

Also added a small tweak to the 'update all' tool which will fix black faces


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@14175 954f8c5b-7b00-dc11-b283-0030488c597c