Commit patch #2439: Mesh replacement in BGE will react properly to armature deform
Changing the mesh of an object that has a deform controller (armature) is now properly handled. The new mesh must have vertex groups matching the armature bones. In simple terms, the new mesh must deform correctly when you assign it to the object in Blender and you test the action. It will deform the same when you replace the object mesh during the game.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@14335 954f8c5b-7b00-dc11-b283-0030488c597c