Bugfix #8948: Transformation constraint - destination rotation is in radians
Unfortunately, my previous fixes here to get gears working resulted in this problem. The flip side of that was that gears worked for quite a few cases. 

Alas, it is not to be... The bug is now fixed, but gears no longer work any more. Sadly, there is not much I can do about this.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@14435 954f8c5b-7b00-dc11-b283-0030488c597c