Surface bugfix (reported by jms on the french newsgroup)
Adding Donut or Sphere surfaces with the Align to View option turned on added broken surfaces.

Those two surface type were always aligned in the past and their creation method were not made to deal with this. Fixed now (that is, they are aligned to view and created correctly if the option is on).

There might be other cases of this bug elsewhere.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@14645 954f8c5b-7b00-dc11-b283-0030488c597c