Todo [#5743] Rotate dosnt work at high zoom
Fixed by only conditionally resetting the start of the triangle used to calculate the angle to when an angle is actually measured (so really small angle don't result in continuous deltas of 0 degree).

However, the smallest rotation angle is still limited by the precision of the acos function (here: 0.02 degrees, 0.02 / 30 with Shift pressed).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@14828 954f8c5b-7b00-dc11-b283-0030488c597c