* Armature hierarchy selection tools These are for use in pose mode or armature edit mode, to let you quickly traverse up and down a chain of bones. It's quite useful for bones that are in hard-to-click places.
The tools are:
*Select parent/child ( [ and ] )
selects the parent or child of the active bone, deselecting the original active bone

* Extend select parent/child (shift [ and shift ] )
selects the parent or child of the active bone, adding to the selection

Thanks to Joshua for reviewing this so promptly!

PS. I'd like to use these [ and ] keys more widely in blender as consistent 'select next / previous' tools. I can imagine it being very useful for a lot of things like keyframes, nodes, mesh edges, etc.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@15944 954f8c5b-7b00-dc11-b283-0030488c597c