Fix segmentation fault when try separate a mesh with the active face selected.
If the active face is part of the selection, Blender crash when
split the mesh because the act_face have an invalid pointer.

The solution is simple, check if the active face is part of the selection
and in that case, set it to NULL.

Campbell please double check this.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@16337 954f8c5b-7b00-dc11-b283-0030488c597c