Grease Pencil - Available in Image Editor:
Grease Pencil is now available in the image editor. It is important to note that the strokes drawn WILL NOT become part of the image visible at the time.

Unfortunately, 'fancy' stroke drawing cannot be enabled for use with the 'Stick to View' setting here, as the scaling is wrong.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@16376 954f8c5b-7b00-dc11-b283-0030488c597c