Python API ---------- Add access to MTex objects from Lamps and Worlds (first pass).  Since the MTex structure is slightly difference between materials, lamps, and worlds, a field is added to the BPy MTex object to distinquish which type it wraps.
Attempting to access attributes which are unique to materials for lamp or
world MTex objects throw an exception.  The next pass will implement MTex
attributes which are specific to Lamps and Worlds.

A new attribute (textures) is added to each module.  It is compatible with
the previous Material.getTextures(), which returns a tuple of either MTex
objects or None.  Surprised we never added an attribute for this before in
all the changes and refactoring.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@16427 954f8c5b-7b00-dc11-b283-0030488c597c