Bugfix #11712
Definitely one of the oldest bugs ever (1995 or so). 
Case is a path (child on path, or deformer, or motion modifier) where the
child is far away from path (300 units or so). In that case you can see
the path jumping to another position a bit after a few frames.

Reason:
For interpolating path positions, I was using bspline code still having a 
very ancient constant 0.1666f. 
Floats have higher precision, like 0.16666666. That solved it :)



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@16994 954f8c5b-7b00-dc11-b283-0030488c597c