== Python: Space Handler Scriptlinks ==
Bugfix (patch #18216)

I made a mistake on my previous try to fix this. I fixed something, tested, decided I should move the call to run spacehandlers to a better place, but missed removing the old call (ugh). Thanks Steven Truppe (rocketmagnet) for emailing me about it and for the patch.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@18671 954f8c5b-7b00-dc11-b283-0030488c597c