[#18257] Workaround for bug in freealut 1.1.0 by Jörg Müller (nexyon)
Since this area is not maintained, I can only say it works on my system (linux/32bit)
And nexyon assured me he checked over this patch well.
Blender using the deprecated function is not great, but nexyon's bugfix would not be available until the next version of freealut so better to work around it by using the new function.

Please test game sounds play back on win32 and Mac, in cases that it worked in the first place.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@18834 954f8c5b-7b00-dc11-b283-0030488c597c