BGE Python API
Use each types dictionary to store attributes PyAttributeDef's so it uses pythons hash lookup (which it was already doing for methods) rather then doing a string lookup on the array each time.

This also means attributes can be found in the type without having to do a dir() on the instance.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@19579 954f8c5b-7b00-dc11-b283-0030488c597c