update for DXF-Importer v1.12, DXF-Exporter v1.29 - 2009.04.11 by migius  - added DWG support, adapted Stani Michiels idea for binding an extern DXF-DWG-converter
The external DXF-DWG converter "dconvertcon.exe" 355kb comes from openDesignAlliance www.opendwg.org, is free, but not opensource, not GPL-compatible (can not be bundled with commercial programs), so must be downloaded and installed in Blender script folder by user. It is a DOS console application, can be started in background mode, works on Windows, and in Wine on Linux and OSX. Version 1.0 (2002) supports conversion between DXF<->DWG for autocad release: 2.5, 2.6, 9, 10, 11, 12, 13, 14, 2000, 2002. 


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@19722 954f8c5b-7b00-dc11-b283-0030488c597c