patch [#18950] Get/set bone and type of drivers by Alberto Torres Ruiz (dithi)
--- quoting the patch submission
This patchs adds the properties IpoCurve.driverBone and IpoCurve.driverBone2 and modifies IpoCurve.driverChannel to
allow OB_ROT_DIFF.

It sets the driver type to pose if IpoCurve.driverBone is not empty or None. Otherwise the driver type is set to
object.

Attached is the patch (with python doc) and an example .blend.

It also fixes the confusing description of IpoCurve.driver



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@21223 954f8c5b-7b00-dc11-b283-0030488c597c