As discussed with Campbell on IRC:
Made some UI operators not register themselves in the console. This made macro creation a pain because the operator list would have tons of splits and other UI commends mixed in with actual data manipulation.

Moved Repeat Last from Ctrl-R to Shift-R so it doesn't conflict with loop cut, whenever it is added.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23157 954f8c5b-7b00-dc11-b283-0030488c597c