More automatic stuff.
Server can now be set to broadcast on local network (every 10s, approximately 20 bytes of data) where client and slave can pick up its existence. This is on by default.

Default ip address is now "[default]", which means for the master that it will listen to all interface and for the clients and slave that they will automatically work out the master's address from its broadcast.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23261 954f8c5b-7b00-dc11-b283-0030488c597c