2.5 - Animation Utility Function
Added a utility function to check which transforms for an object or bone are animated, returning these as bitflags and/or optionally retrieving the relevant F-Curves too. Beware that this method may not be working correctly yet, but it shouldn't hurt anyone in the meantime :)

Also, split RNA-path building function up into a version which only creates the path up to the given struct, with the other parts being added later.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23298 954f8c5b-7b00-dc11-b283-0030488c597c