Bone Selections: Ability to set bones as unselectable
In the Outliner, it is now possible to toggle per bone the selectability of the bone in the viewport, as for Objects using the restriction columns. This can also be set using the RNA-api.

I've tested all commonly used tools IMO, but there may still be a few which I've missed. Please report those cases.

PS. For some reason, the define was already there, but not connected up to anything. Can't remember why anymore, but here it is...

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23658 954f8c5b-7b00-dc11-b283-0030488c597c