netrender
Support for fluid files and better support for point cache (including external cache for particles)

This also fixes a couple of bugs with frame based dependencies and with file transfer.

NOTE: With external point cache and fluids, the path needs to be relative or relative to the file (starting with //) if the files are not on a shared drive. It should eventually warn if that is not the case, but doesn't right now, so be careful.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23727 954f8c5b-7b00-dc11-b283-0030488c597c