Project option for snap to faces.
This is similar to the old retopo all option but uses the snapping code and not the openGL depth buffer (it's thus more precise).

Not sure if making it available as a snap option is sensible, this is up for discussion.

NOTE: it will get slow fast on large meshes, we need to plug in an acceleration structure into snapping.

This will need an icon too.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23798 954f8c5b-7b00-dc11-b283-0030488c597c