Armature Ghosting Bugfix:
I noticed that armature drawing seemed to be using the wrong colours for ghosting in some situations (in particular, Cessen's biped rig for 2.5). Commented out what appears to be the offending code for now. 

Not totally sure why this broke yet, but as possible bonus of this, we sometimes get coloured ghosted controls (if bone groups were in use). 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23808 954f8c5b-7b00-dc11-b283-0030488c597c