Num Sliders Tweak:
The dark bar for sliders will now be drawn in a way that takes into account the state of the slider (this includes info about keyframing and/or drivers). This means no more ambiguarity about whether these sliders are animated or not!

I've had to make a little hack here to make sure that the sliders bar will still be visible after the state has been blended, but this should probably be solved nicer by adding a separate setting for blending-factor in this case.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23860 954f8c5b-7b00-dc11-b283-0030488c597c