Patch #19034: (2.5) Display titles of closed panels in horizontal mode
This patch was one of the patches submitted by Wolfgang W. (bender) (the others have already been committed). In the patch, there was commented out code for counter-clockwise rotation of the text, but I've removed that.

--

In this commit, I've also removed all of the panel tabbing code instead of merely commenting it out.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@23933 954f8c5b-7b00-dc11-b283-0030488c597c