* Partial fix for bug [#19734] N panel's transform doesn't display the units
Converted the Object transform properties to use rna/rna buttons. This has the advantage of not only displaying/editing units, but also RMB menu, keyframing, drivers, etc too. Part of this was to convert 'Dimensions' to an rna property, converting to and from scale. This also allows you to set the object's dimensions via fcurves or python, but note that it's driving the object scale setting internally so if you animate both dimension and scale at the same time one will override the other (i don't expect many people to attempt this).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@24333 954f8c5b-7b00-dc11-b283-0030488c597c