* New tool - Join as Shapes
Available in object mode (Object -> Join as Shapes), only works for meshes at 
the present. Will merge all selected objects as shape keys on the active object, 
if the vertex count is the same.

This does not keep references to the external objects like in some applications, 
rather it's a quick way to update the shapes on the active object (perhaps after 
importing new versions from external applications).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@24960 954f8c5b-7b00-dc11-b283-0030488c597c