Transform Modal Map:
Adding events for X/Y/Z axis and planar constraints (always use the user orientation, not global first) and for turning constraints off.

Built-in keys (toggle between global, user and off) are not in the map but can be overridden (any events handled by the modal keymap will not enter the built-in map).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@25037 954f8c5b-7b00-dc11-b283-0030488c597c