Sequencer: Transform Strip updates (Durian wish)
Removed the Start/End Settings and the effect_fader from the sequencer transform strip. 
The transform strip can now only be animated directly by keying the Scale X, Scale Y, Translate X, Translate Y and Rotation values.

Caveat:
The uniform scale re-uses the Scale X value, so when changing the animation on Scale X in the non-uniform scaling case, the uniform scaling will be affected too. This was done to not break files for Durian. 

Note:
As much as I would have liked to clean up the TransformVars, it caused crashes of earlier versions of blender when opening the file.
The rna identifiers were also kept as to not break any existing animation on those keys.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@25326 954f8c5b-7b00-dc11-b283-0030488c597c