Grease Pencil: Lock current frame option
It is now possible to make Grease Pencil Layers to keep displaying and editing the current sketch-frame with this option. This allows to draw a frame which contains markings made for different times (i.e. a spacing/timing chart that you can keep adding to as you scrub to different points on the timeline).

Use the clipboard/camera toggle (the one beside the visibility toggle) to enable. This should get an icon of its own at some point...

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@25820 954f8c5b-7b00-dc11-b283-0030488c597c