netrender:
thumbnails on the web interface (resized with imagemagick or compatible if present, fullsize otherwise)
Clicking on the "show" link shows the thumbnail associated with one frame. Clicking on a second one shows all frames in the range. Clicking on more frames extends the range. Clicking on a thumbnail (or on the show link of a visible thumbnail) hides all thumbnails.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@25822 954f8c5b-7b00-dc11-b283-0030488c597c