BGE API_DOC: VideoTexture and PhysicsConstraints templates (need the functions to be filled with their description).
PhysicsConstraints is documented in the Game Kit Book:
http://download.blender.org/documentation/gamekit1/

VideoTexture is documented in the wiki:
http://wiki.blender.org/index.php/Dev:Source/GameEngine/2.49/VideoTexture

I don't think I will have time to fill the documentation. But I hope this commit helps someone interested in helping it.

Therefore volunteers to document those modules are highly welcome !!! (let's give to BGE the documentation it deserves)!

* + added GameLogic.Lave/LoadGlobalDict + some typo fixes

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@25875 954f8c5b-7b00-dc11-b283-0030488c597c