Text Editor: Add an option "Tabs as Spaces". So now tab is not ALWAYS converted to spaces. This is stored by text datablock (what allows to do nice things in the future, as automatic check for the indentation type of the file).
Ideally we should redraw the other Text Editor windows after changing that (in case the same file is opened and the Property panel is also open). Not sure how to do that though.

I'm using TABSTOSPACES as the DEFINE flag because TABSASSPACES sounds too ugly.

(also fix for interface divisor bug)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@26007 954f8c5b-7b00-dc11-b283-0030488c597c