DopeSheet Keyframe/NLA Strip Scaling Tool:
Scaling behaviour relative to mouse movement was inverted. Made the this transform tool now use the standard input mechanism for scaling inputs so that the behaviour is now correct. 

Martin: In my tests, the responsiveness of the scale tool is now not sensitive enough though. If you have some time, could you look into whether this can be made to work a bit freer?

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@26064 954f8c5b-7b00-dc11-b283-0030488c597c