Animation Editors: DopeSheet filtering option to only include Objects that are members of the specified group
Between the datablock filtering options and the auto-snapping menu in the headers of Animation Editors, there is a new toggle for enabling this new feature (only when there are groups in the scene). Enabling this, a field to enter/select a group in the scene to filter with, will appear beside it. 

This feature has been added to make it easier to manage animating multiple character shots in Durian. 
By assigning the rigs of several bandits to a single group, or Sintel and her staff to another group, or Sintel and the guardians to another group, and so on, it is possible to filter the animation data shown in the editors to a useful subset of the total motion in the scene. 

This makes it easier to retime or edit the motions of one set of characters and their props without affecting the motions of other no-related objects. The downside is that there is a bit more setup work required upfront, but that's probably a small price to pay for some groupings that may be useful in other ways too later (perhaps for compositing or lighting work).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@26205 954f8c5b-7b00-dc11-b283-0030488c597c