Fix #20327: uv smart project options can't be tweaked.
Python operators calling C operators would get too many undo pushes,
causing redo of the python operator not to work. Now the depth of
operator callbacks is counted to detected nested calls, and in that
case skip the undo/register here, and only do a single undo/register
for the mother operator.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@26395 954f8c5b-7b00-dc11-b283-0030488c597c