Bugfix #20935: Evaluation Time For Curves No Longer Accepts Input Over 1.0
Reverting the part of an earlier commit to show the Evaluation time in the Path panel that made the Evaluation Time setting a factor. This setting should not be a factor, since it gets divided by 'Path Length' to normalise it to the 0-1 range needed. When this setting isn't animated, the evaluation time setting is automatically set to the current frame number, so that when divided by Path Length, child objects still follow the curve. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@26570 954f8c5b-7b00-dc11-b283-0030488c597c