Bugfix #21384: Bone Driven Shapekeys Child Evaluation Problem
Transform channel drivers for bones in 'localspace' was using the wrong matrix when getting the transforms. I had been assuming that pchan->chan_mat always contained only the matrix-ised transform values stored in the pchan (which is true while constraints are being evaluated, but not afterwards). 

Changes:
- Added a new function to calculate this matrix instead of directly writing it on the pchan->chan_matrix field. 
- Also, made the normalisation of the quaternion values during this process be done on a temp var instead of on the stored value. This was a constant source of confusion in the past, so let's see if we can do without it now :)

Unrelated to this commit, I've also fixed a compiler warning with previous commit that I missed (missing include).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@27200 954f8c5b-7b00-dc11-b283-0030488c597c