python api function for rna objects: object.as_pointer()
This means we can write low level apis in pyton or C where blender data is passed to external C modules without having to have blender support this directly.

Example use case is to get an image pointer then use ctypes to get the image buffer and pass it to a C image processing function.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@27292 954f8c5b-7b00-dc11-b283-0030488c597c