Patch [#21750] Add luma waveform and vectorscope to image view by Xavier Thomas
This adds the waveform monitor and vectorscope to the image editor 'scopes' 
region, bringing it inline (plus a bit more) with sequence editor functionality,
and a big step closer to the end goal of unifying the display code for image/
comp/sequence editor. It's non-intrusive, using the same code paths as 
the histogram.

There's still room for more tweaks - I modified the original patch, changing 
the openGL immediate mode drawing of the waveform display to vertex arrays for 
speed optimisation. Xavier can look at doing this for the vectorscope now too.

Thanks very much Xavier!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@28027 954f8c5b-7b00-dc11-b283-0030488c597c