Fix [#21890] YUV->RGB: Color clamping 16-235 in all motion pictures (ffmpeg)
Patch by Troy James Sobotka - this uses options in newer FFMPEG versions to 
convert the full 0-255 YUV range of imported imagery to RGB, rather than 
clipping at 16-235.

This functionality is not available yet in an official FFMPEG release 
(current precompiled version in /lib for osx at least is v0.5.1 from 2009) so this 
won't take effect in that situation, but if you've got a newer ffmpeg on your system
it will work.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@28084 954f8c5b-7b00-dc11-b283-0030488c597c