fix for [#22195] Empty Size minimum limit too big (patch by Bassam Kurdali - slikdigit) with small change (ui range precision to 2 instead of 4, since 4 doesn't seem to work here).
From the tracker:
""" 
The smallest size allowed for empty_draw_size is 0.1. This is un-necessarily limiting and, depending on the scene, much
too large resulting in visual confusion in the 3D View.
Opening old files initially results in small empties, but when you click on them they 'grow' to the new minimum. This
seems to show that previous blender versions didn't have this bug.
Please remove the limitation, or, if needed, make it a much smaller number, so that empties are easier to see and distinguish from one another in small scenes/setups
"""

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@28498 954f8c5b-7b00-dc11-b283-0030488c597c