Add the option to show also the UVs of other selected objects in image space / uv edit. The code was already there, and the option as a rna bool, but no ui to set it. Matt figured that the View menu in image space next to other uv stuff, which only shows when UVs are edited, is the right place.
Works so that when entering editmode for an object to edit UVs, when 
have also other objects selected and this option on, also the UVs of 
those other objects are shown in the image view.

Liquidape asked this on IRC, and we thought the feat doesn't exist, so I 
looked out of curiosity in the code as was thinking it would be easy to 
add. Was surprised to find it there already :)

First time that did anything with 2.5, was sure fun enough to search 
thru the code to figure out how things work. Adding this ui thing proved 
to be exactly as trivial and nice as it should, and the things under the 
hood seemed nice, yay!



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@28632 954f8c5b-7b00-dc11-b283-0030488c597c