Fix for compilation with newer ffmpeg library
Error was cased by undefined UINT64_C in stdint.h for c++
programs. As I understand from ffmpeg mailing list the simpliest
solution is to define __STDC_CONSTANT_MACROS before including
ffmpeg libraries. There already was some amout of such definitions in
the code and I've tested this with newer and older ffmpeg
libraries -- this should work and should be safe, but maybe
there is better way of fixing this problem.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@28793 954f8c5b-7b00-dc11-b283-0030488c597c