Fix #22534 linking object data for two forces results in crash.
This is because the make_links_data_exec don't check for the
object type like before, so try to access the obdata of an empty
and blender crash.

The solution is not the best, we have a new function
allow_make_links_data to check if we can links data from one
object to another.

The real solution is build the menu like the 2.4x, so only
show the options that we allow for the object type that
we have select/active.

Matt, any suggestion ?

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@29346 954f8c5b-7b00-dc11-b283-0030488c597c