Revert revision 29735:
Fix #22051: crash when scaling parent metaball. Keep the constant resolution
for any motherball's scale.

This avoids running out of memory when scaling the metaball down, but there's
a reason it depends on this scaling, for example for instancing it's more
useful to have this. It also doesn't really solve the problem but only moves it,
it's still possible to run out of memory with different setups/scales.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@30031 954f8c5b-7b00-dc11-b283-0030488c597c