Bring back "Sort Faces" command.
This was the old Ctrl + FKEY in object mode, now it's in
edit mode and is part of the Ctrl + FKEY menu (Faces).

I also assing this to the Ctrl + Alt + FKEY, but Matt please
check this and feel free to change (or tell me and I will change).

Still there is no "reverse" function, but I commit now to finish
in my home.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@30056 954f8c5b-7b00-dc11-b283-0030488c597c