Fix for #22818: blender doesn't find a systemwide installed python.
What happens is that blender looks for a directory "python" in the same
place as the executable for local installations, but that also means when
you have /usr/bin/blender it will look for /usr/bin/python, which is an
executable. Now it checks if it is actually a directory and not a file.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@30310 954f8c5b-7b00-dc11-b283-0030488c597c