[#23032] Bracket Highlighting in Text Space Fix [Patch to fix attached] from Justin Dailey (dail)
 from the tracker
 --- snip ---

In the text editor doing something like this:

print(":(")

When it goes to match the closing bracket, it will highlight the one in the string, not the first one. Also doing:

array["[index"]

will cause it to match the second [ with the closing one.

I have attached a patch to fix this issue. (See attached image to see correct highlighting)
It also works with triple quotes strings(ie """...""" or '''...''')

*Note* However, originally bracket highlighting always on even if syntax highlighting is off. The patch makes it so
it only highlights brackets when syntax highlighting is on (this is a side effect of doing the code this way, if it
was done any other way ALOT of code would have been needed to check for strings,triple quoted strings, escaped quotes,
and comments forwards and backwards). When highlighting matching brackets, the code checks the line's format string
to see if the char is in a string or comment to skip it. If syntax highlighting is turned off, the format string is
null and cannot be used,thus no bracket highlighting.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@30764 954f8c5b-7b00-dc11-b283-0030488c597c