patch [#23041] CMake WITH_PYTHON_INSTALL option from Dan Eicher (dna)
--- snip ---
Adds the option to not copy python into the .blender/python directory.

Also fixes the problem of deleting the blender installed python where the blender runtime only looks for the copied python and not the system one but not entirely sure why, something to do with CMake hardcoding the PYTHONPATH and PYTHONHOME vars I suppose.

This patch only applies to the 'UNIX AND NOT APPLE' platforms but, then again, those are the only ones anyone really
cares about...or at least the only ones where using the system python is really viable.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@30779 954f8c5b-7b00-dc11-b283-0030488c597c