Fix for Mesh.uv_textures.new(name="my_uv") returning the wrong uvmap - reported by Vitor Balbio - not in tracker.
The code was taking the last layer, but that is only valid if the mesh has only one kind of CustomData types (e.g. only UVMaps or only VertexColors). The solution I found is to call CustomData_get_named_layer_index instead. To avoid some situations where an uv with this name may already exist and the number of UVs is already the limit we are returning a CDL only when the texture is properly created.

As a bonus that also fixes the same problem with VertexColor.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@31601 954f8c5b-7b00-dc11-b283-0030488c597c