Fix [#23613] Align active camera to view vs toggle camera acts odd Reported by Reiner Prokein
After aligning active camera toggling between last view and camera worked wrong in such a way that the view previous to the one to which was aligned was rotated to. In some cases rotation to last view seemed incomplete and ortho was used. Thanks to M.G. Kishalmi (lmg) for showing through a Skype screen share.

This is now fixed by copying the current view info before active camera is aligned.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@31712 954f8c5b-7b00-dc11-b283-0030488c597c