Fix [#23791] Crash rendering to some format Reported by L Harch
Apparently the FFMPEG libs were the problem here. Updating to a version from April 22nd, 2010. I chose this version, since it's also the last day a win64 build of the development packages is available, so we'll have the same revision in use for both 32bit Windows and 64bit Windows.

This revision probably will fix a lot of open bugs, I'll be reviewing those FFMPEG issues after this commit and mark them in the tracker if I can't reproduce them anymore on win32 nor win64.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@31909 954f8c5b-7b00-dc11-b283-0030488c597c