Bugfix #24099: nla content moves out of sync - with fix
Thanks Shane Ambler (sambler) for the patch!

<quote>
This kind of follows on from a report that I think was closed prematurely - #22775 - I can't find any reference to an
outstanding todo that relates to it.

When resizing the nla editor the channel names and the main area get out of sync.
When toggling back from fullscreen the content is hidden off the top of the area requiring scrolling to see it.

The dopesheet displays similar problems but after fixing the ui_view2d_sync call it appears to behave as if the v2d.keepofs
has been set for the most part.

Two areas seem to be related to this - the first is calls to UI_view2d_sync used for these two views use the wrong
flags.

The other is v2d.keepofs not being set. (dopesheet has less issue here but I think it is meant to be set the same as
nla)

</quote>

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@32509 954f8c5b-7b00-dc11-b283-0030488c597c