Partial fix for #24773: Material Nodes - there isn't able to set keys on Mapping coordinates
Playback now works. The problem was that material/texture nodetrees were getting ignored completely, as I was assuming that all of these existed in the main->nodetree collection when in fact only the "group" nodetrees lived there. I don't really agree with this, but that's the way it is...

TODO: animation editor support still to come

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33247 954f8c5b-7b00-dc11-b283-0030488c597c