Fix #24834: curves extrude + bevel gave bad normals on rendering.
This happened when removing vertex normal flipping in the render engine. But
actually the normal of the first face in the curve was flipped on rendering,
because it used a different vertex order than other places. Also removed the
hack that computes a face normal only from a triangle instead of a quad.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33269 954f8c5b-7b00-dc11-b283-0030488c597c