Bugfix #25049: Compositing Nodes not Keyframable
An error seems to have been introduced to the node-tree building at some point, which means that the ID-type for data-attached node trees was incorrect (i.e. scene->nodetree->id.name = NTREE_COMPOSIT instead of ID_NT). 

This in turn meant that the ID AnimData availability poll would fail, as the ID-type could not be determined.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33534 954f8c5b-7b00-dc11-b283-0030488c597c