Bugfix #24163a: Unable to animate INSIDE a group node in the compositor (This commit doesn't fix the original bug reported in the report, but does fix one that was discovered while investigating that one)
Trying to insert keyframes for nodes within group nodes was failing. 

This was caused by the ID-block for those UI widgets being set to the Node Editor's current ID-block (i.e. NodeTree) vs the NodeTree that those nodes lived in. The net result was that the paths couldn't be resolved, as the paths obtained for those widgets could only work up to the group's nodetree.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33535 954f8c5b-7b00-dc11-b283-0030488c597c