Bug report #25384
Node editor
Some nodes - when muted - didn't draw a red line to indicate
a pass-on action. Stil, the node then is muted and either passes
on a buffer of same type, or just black.

To more clearly indicate muting, it now draws the header in a 
reddish blended color too.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33900 954f8c5b-7b00-dc11-b283-0030488c597c