Bugfix, irc report.
UI name buttons with Search callback show red when item couldn't be
found. The check was too simple though, since the searches also find
partial matches. This caused KeyingSet browse to show red for "RotScale".

Now 10 matches are stored, and each match is being checked. Still a
bit weak for cases with more than 10 matches, but in that case it
won't show red-alert.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@33927 954f8c5b-7b00-dc11-b283-0030488c597c