Constraints UI Tweak:
Experimental tweak for the "Object Constraints" tab so that it now
shows a warning message instead of the "Add Constraints" button when
the active object is in Pose Mode. Hopefully this will further
alleviate any confusion over Object vs Bone level constraints.

Todo:
While coding this, I noticed that we currently don't have any way of
making help-text labels in UI panels which can span multiple lines
(word-wrapped or manually split). Probably not a critical issue, but
it would be nice for completeness...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34017 954f8c5b-7b00-dc11-b283-0030488c597c