Marker Bugfixes ("Make Links to Scene" operator):
Some tweaks for this operator, found while scanning through the
current state of the markers code. These seem to be correct judging
from the overall direction the code is going in, though the intention
of this operator seems a bit unclear from the naming.

- Fixed potential issue where with only some markers selected (and the
last one selected not being the last marker in the source scene), you
could end up chaining the source and destination marker lists together
improperly.
- Clarified description/error messages...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34093 954f8c5b-7b00-dc11-b283-0030488c597c