Ghost:X11 Set the default max width and max height value.
Some window manager can set default value of this to be the
screen size, so running blender with -p or --window-geometry
don't work with value bigger than that.

This commit try to "avoid" the bug #25709, but at the end
depend on the window manager, so maybe work or maybe not.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34420 954f8c5b-7b00-dc11-b283-0030488c597c