Bugfix [#25737] Console error message
It seems that rna_def_effect() is missing a rna_def_input(), resulting
in:

  rna_uiItemR: property not found:
EffectSequence.animation_offset_start
  rna_uiItemR: property not found: EffectSequence.animation_offset_end

in the console. I'm not sure whether these settings are used for
"effect" strips or not, but I'll leave that to sequencer guys to
figure out if/when they get a report about this.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34426 954f8c5b-7b00-dc11-b283-0030488c597c