New Keyframe Type: "Jitter"
While doing some animation work over the past few days, I found that
it would be useful to be able to tag keyframes as a another type yet:
"jitter" for fill-in secondary motion (i.e. blinks/hand-keyed camera
shake). Of course, what you really use these for is up to you, but
this is one more option when you need it :)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34600 954f8c5b-7b00-dc11-b283-0030488c597c