Moved Auto-Keying "Insert for KeyingSet only" option from UserPrefs to Scene (Toolsettings, i.e. alongside "layered" option for using NLA while doing auto-keying)
This option makes all Auto-Keying operations use the active Keying Set
to carry out keyframing operations instead of picking and choosing
their own Keying Sets to use, thus cutting down on the number of
unwanted keys.

Warning: if the older userpref option was enabled in an old
startup.blend, it may be difficult to turn this option off.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34634 954f8c5b-7b00-dc11-b283-0030488c597c