Bugfix [#26002] Outliner Visible layers + Single object & data = broke conections
It seems that some of the Outliner hacks used while building the tree
was causing problems, as Make Single User (and potentially other code
working with ID-data, specifically with the "newid" value there) was
making use of the variable used there for other purposes, leading to
memory corruption.

This bug also occurred in 2.4x, though when I tested there, it crashed
immediately.

Ton, you may want to double-check this bug!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34755 954f8c5b-7b00-dc11-b283-0030488c597c