DopeSheet: Add menu entries and hotkeys to select all keyframes left/right to current frame in dope sheet
This commit separates out this selection functionality out of the
click-selection operator into a separate operator, so that hotkeys and
menu entries can be assigned to it.

This is based on an idea+patch (#23738) submitted by Torsten Rupp
(rupp), though I've ultimately decided not to go with the suggested
implementation as I don't think this fits that well under the "column"
select operator.

Todo: Graph Editor support will be coming shortly...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34831 954f8c5b-7b00-dc11-b283-0030488c597c