Commit patch [#25939] material panel proposal by Ervin Weber (lusque). Thanks!
From the patch description:
"A new panel is proposed to bring togheter all the properties of a material that belong to the render pipeline level.
Such properties are currently not mixable with node materials, as nodes operate on a shader level."

Commiting this patch as approved in the sundy meeting. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@34914 954f8c5b-7b00-dc11-b283-0030488c597c