Py/RNA Stability: don't allow python to reference freed ID's and crash. Second method for not having python crash blender on invalid access (ifdef'd out ATM, so no functional change).
This uses a weakref list per ID, and invalidates all members of that list when the ID is freed.
the list is not stores in the ID pointer but using a hash table since storing python in DNA data is not acceptable.

This is more correct then the previous method but shows down execution of scripts significantly since its always adding and removing from lists when data is created and freed.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35282 954f8c5b-7b00-dc11-b283-0030488c597c