Py/RNA api - real collection iterator Previously the api just converted the collection to a list and got the iterator from the list to return.
This has the advantage that it uses minimal memory on large collections where before it would make an array.

Though the main reason for this change is to support a bugfix for collections which free memory when they are done, this currently crashes the python api since once the list is built, the data is freed which is used by the list items in some cases (dynamic enums for eg).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35322 954f8c5b-7b00-dc11-b283-0030488c597c