Bugfix #26442
Very bad crashing in using "Not AA fonts" and drawing length info
on EditMesh. This uses the call BLF_draw_ascii() which apparently
corruptes then badly. Disabled the call for now, and wait for fix.

In the course of testing, added proper strlen storage for draw
in this code.

Added strlen



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35452 954f8c5b-7b00-dc11-b283-0030488c597c