Bugfix #26444
Double click in File Window also loaded a file on double-click at
a directory.

The operator for it was using a Macro mistakingly here. On any double
click, it selected the item first and then executed load.

By default, any double click action has to rely on the first click 
being properly handled before. Simply removing this macro, and assign
the "exec operator" to the double-click event works as expected.
A double click on a directory then just shows the dir (because the exec
requires an 'active' file item).



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35473 954f8c5b-7b00-dc11-b283-0030488c597c