Bugfix: [#25725] Transform bone constraint & 'local with parent' as owners space [#26014] ChildOf Bone Constrain (influence) works in wrong workspace
Moved the influence calculation stuff outside of the space conversions
(i.e. so that it is done in worldspace only) fixes these problems,
which seem to arise when a constraint doesn't work in worldspace AND
doesn't need to apply inverse correct for this space conversion when
it's done, hence resulting in mismatch between spaces for old and new
matrices resulting in all the weird behaviour.

Patch to fix this from Jahka. Cheers!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35759 954f8c5b-7b00-dc11-b283-0030488c597c