Apply patch [#26660] Scons build on OSX fails to create Info.plist file fix provided by Mike Sloman
from the patch report:

When building on Mac OSX with SCONS the Info.plist file is zero length and thus OSX cannot identify icons and file type associations.

The problems was in the execution of the "date" command to add times in the modified bundle name strings passed to sed. This failed and thus the info.plist file was overwritten with 0 length file.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@35893 954f8c5b-7b00-dc11-b283-0030488c597c