Apply console part of patch [#26044] Windows thumbnails and improved filetype registration submitted by Tom Edwards
This patch introduces a switch -con and its longer version --start-console. When giving this on cmd-line you'll get the black console window. The new behaviour is to hide it by default. We'll still see briefly the console at startup and during exit, but that's something that cannot be changed.

If you start blender from a cmd.exe, the console will not be hidden.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@36057 954f8c5b-7b00-dc11-b283-0030488c597c