View Dolly Operator: Ctrl+Shift+MMB.
Recently there were 2 reports about zoom not working right because at some point you can't zoom in any further.

This is not actually a bug with zoom but a limitation in blender that there is no way to move the viewpoint forward (unless you count rotate 90d, pan, rotate back which is a crummy workaround).

So adding view dolly operator:
- Supports zoom to mouse position setting.
- Supports dolly hoz/vert, invert setting.
- Moves by a fraction of the view 'dist', so the zoom distance my be roughly in proportion to the scale of objects in the scene.

only used in perspective view since this is not useful in camera/ortho view.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@36283 954f8c5b-7b00-dc11-b283-0030488c597c