Bugfix [#27452] Wrong linking of character groups
Linking in groups also ended up adding those objects to the current
scene that we're linking the group in to. This is a regression. It
leads to more work (and/or confusion) from animators when just trying
to get a character into a shot file to start animating it (see my
notes on my blog for the proper workflow regarding this).

Reverting r36762, which caused this mess.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@36808 954f8c5b-7b00-dc11-b283-0030488c597c