bugfix for: [#26753] PhysicsConstraints ID trouble on 64bit (linux at least). [the problem also affected OSX]
PhysicsId are Long, not ints (see PyObject* KX_GameObject::PyGetPhysicsId() )

There is a reference in the code to use PyCapsule instead of int. I'm not sure
about that. This patch at least stops the crashes
(update: I talked with Campbell and he repeated that PyCapsule are better, but if long is working it's fine for now).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@36970 954f8c5b-7b00-dc11-b283-0030488c597c