Fix #27505: Text Editor always indent next line when a " is found (which is not always correct)
Do not indent if there's any non-space character after colon.

This only makes life a bit easier, but it's still not 100% correct indentation
strategy. For example when colon is inside non-closed string or so.
Also there's not indentation for { and un-indentation for }.

Handling such cases would require much smarter strategy..

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@36983 954f8c5b-7b00-dc11-b283-0030488c597c