Bugfix #27768
On clicking in a non-active Blender window (when you activated others),
the mouse position of the first click was still the old position.

Problem is in GHOST; it sends out the 'activate window' event after the
mouseclick event itself. Code now checks for this case and reads the
correct mouse position.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@37854 954f8c5b-7b00-dc11-b283-0030488c597c