Bokeh blur in the blur node is wronlgy calculated.
when using the node on a single white pixel on black background, the output should look like as the bokeh image.
being a round image, but it looked like a donut.

the make_gausstab used dist/rad and bokeh used (dist/rad)*2 - 1
I changed it to reflect the correct bokeh circular image
 



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@38347 954f8c5b-7b00-dc11-b283-0030488c597c