Simple argument to be able to explicitely tell the bitness you want to build Blender in.
python scons\scons.py BF_BITNESS=32
python scons\scons.py BF_BITNESS=64

So from now on for Windows you don't have to run a specific win32 or win64 Python version
to get the Blender version you want. If you omit the BF_BITNESS flag the bitness as per
Python version will be used as before.

Note that this is an argument, so works *only* on the command-line, not in your user-config.py

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@38953 954f8c5b-7b00-dc11-b283-0030488c597c