cmake files for bundling blenderplayer in OSX this is working fine. The only problem is described in the XXX there.
While Blender is linked to the right place during `make` time (~/bin/blender.app/Content/MacOS/blender)
blenderplayer is in the top bin folder (~/bin/blenderplayer).

That means it will be correct only after you run `make install` (we are moving it to the right place inside the bundle).
I couldn't find the place to fix this properly. I will see with Jens Verwiebe how we can address that.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@40548 954f8c5b-7b00-dc11-b283-0030488c597c