making PO/POT/... update scripts py3.2 compatible + making it almost win compatible
Adding from codecs import open + the encoding in update_pot.py so it works for python 3.2.
Now it should be easier to replace the executable folders in the first lines of the file.
It now also works if you are in the same folder as the .po files (i.e. os.path.dirname(__FILE__) is empty)

It's still not working for me on Mingw because I need gettext 1.8 and so far I found only the 1.7
(in 1.7 msgmerge --lang is not supported)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@40681 954f8c5b-7b00-dc11-b283-0030488c597c