Fix #28914: crash loading file saved with cycles builds in trunk.
The cause is an unknown node socket type in node groups. Ideally the node system
should handle this better and remove the unknown sockets from groups, but this
is a bit of a risky fix to do now, so instead the shader socket type has been
added, since this is a simple change and the code has been tested well.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@41009 954f8c5b-7b00-dc11-b283-0030488c597c