While reviewing this part of the code for a Blender Proceedings article, I noticed that the lock enforcer was taking several improper steps to redistribute evenly with multi-paint.
Beforehand, a few problems were in view: some of the indexing was changed towards the end to use the more efficient stack structure, but needed to use the correct def group indices.

Also, the designated selected group would use its own value to acquire the standard to base change distributed to the others.

Lastly, the total_change was used rather than -left_overs in the formula to compute the new designated weight within the means of the locks' allowed changes.


Now, while maintaining the ratios of the selection, it correctly returns left over change that could not be redistributed to the unlocked groups.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@41711 954f8c5b-7b00-dc11-b283-0030488c597c