regarding [#29521] Invalid Vertex Weights
added checks to mesh.validate() so it checks for nan weights and negative groups which can crash blender.

the file in the report looks to be corrupt, but this at least can make it load.

also noticed validate was using isnan() and finite() were both used to check valid vertex locations, only finite is needed.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@42508 954f8c5b-7b00-dc11-b283-0030488c597c