Fix #29687: Normal bake from multires gives distorted results.
It is just a limitation of multires baker which doesn't deal correct with
baking to subdivision level 0. It was supposed to work with levels on which
sculpt data is affecting on mesh, so interpolation between grids works correct.

Fully accurate baking in this case will need raycasting stuff which will make
it much slower and will remove main benefit of regular baker -- speed and
low memory usage.
Another option would be to make multires apply sculpting data on level 0,
but it's not related at baking at all and has got it's own difficulties.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@42979 954f8c5b-7b00-dc11-b283-0030488c597c