Color management: add "Color Unpremultiply" option for images and render settings. For premultiplied alpha images, this makes any color space conversion for the image or render output work on color without alpha multiplied in.
This is typically useful to avoid fringing when the image was or will be composited
over a light background. If the image will be composited over a black background on
the other hand, leaving this option off will give correct results.

In an ideal world, there should never be any color space conversion on images with
alpha, since it's undefined what to do then, but in practice it's useful to have
this option.

Patch by Troy Sobotka, with changes by me.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@43004 954f8c5b-7b00-dc11-b283-0030488c597c