Bugfix [#29629] UV animation stops working forever if Edit mode entered Thanks Sergey for the patch!
This bug, and many like it would arise because at some point a F-Curve would get
detected as being unable to be resolved (when trying to display its name in an
Animation Editor), so gets tagged as being "disabled". It was assumed that such
channels usually belonged to deleted data or stuff that wouldn't be able to be
resolved again for a long time. However, in cases like UV settings, they may
only be unavailable temporarily.

As a compromise, this patch clears this disabled flag if a channel is found to
be now working when (in the process of finding out its name when redrawing an
Animation Editor) it can be resolved, in the same place where the disabling was
taking place. This doesn't work fullproof in tests I've done, but should be
nicer than before.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@43086 954f8c5b-7b00-dc11-b283-0030488c597c