Fix:   [#25834] no color of textobjects in game engine when combined with 	   textured objects   [#26893] Curruption of displayed text (debug properties/fps info or bgui) 	    when using animated/tile uv mode
The first bug was beacuse a bad mode on the texture environment, now
we save the current glTexEnvi, set the one that we need, draw and
restore the original at the end.

The second was because a missing call to glLoadIdentity for the
texture matrix and as we do before, now we do a gl-Push/Identity/Pop
for this matrix to.

The first problem was solved by Kanttori and the second by Dalai.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@43105 954f8c5b-7b00-dc11-b283-0030488c597c