BGE Font Object: fix for relative path not working AND packed fonts not working [I don't think anyone has ever reported those, what makes me slightly sad but carry on ;)]
Those fixes introduce a more generic function to load a font before calling BLF_load.
I think it should move to be part of Blender util routines or BLF itself.

For the time being here will make it. Once we get <builtin> font working we go for this.
Thanks Diego Borghetti for the usual assistance with blf.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@43151 954f8c5b-7b00-dc11-b283-0030488c597c