Carve: improved handling of meshes with interesting manifolds
Unioning of intersecting manifold tried to perform as little union operations
as possible, but there were some not totally correct assumption which lead to
cases when unioning of manifolds of some mesh might be happened when one of
mesh sets already had got intersecting manifolds.

This commit corrects this incorrect behavior.

Discovered this when was looking into #30175: Boolean Difference causes 2.62 RC1 crash.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@44104 954f8c5b-7b00-dc11-b283-0030488c597c