font objects in blenderplayer fail when object with alpha is rendered the embed bge has its own DisableForText() routine which prepares the GL flags for the text render.
For some reason blenderplayer is taking a slim approach and going straight for the rendering. This routine helps to address this bug/case (non reported, foundin my own project files). If need arises we should move this to its own routine incorporating the other GL tests the embed bge performs.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@44117 954f8c5b-7b00-dc11-b283-0030488c597c