fix [#30266] B-Mesh: Issue with Weight Painting
thanks to Nicholas Bishop for finding the cause of the problem.


don't tesselate on load, this means me->mface will be NULL by default.

we may need to have this set if existing uses of this array are not resolved before release - so add a define USE_TESSFACE_DEFAULT, to change this easily.


this is a rather indirect fix - need to take care here.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@44272 954f8c5b-7b00-dc11-b283-0030488c597c