uv stitch static island highlight ported to bmesh system.
Unfortunately since we are missing a way to tesselate uvs, the result will look good only on concave faces since we are essentially using a triangle fan to draw. When a tesselation of uv polygons is possible the tool should also change to use that instead. This will also influence the preview drawing code, coming next.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@44307 954f8c5b-7b00-dc11-b283-0030488c597c