Fix #30491: Not Updating Scene Length (also fixes special request from Ian for Mango)
Added operator to update actual content length of all selected strips.
Can be useful for scenes and movies as well after doing making changes to
scene/movie.

Can be improved further to deal better with cases when strip has got effect
and it's get reshuffled because of overlapping after changing it's length.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@45083 954f8c5b-7b00-dc11-b283-0030488c597c