Popup menu layout inherits context store from button.
When adding extra context data in a layout using uiLayoutSetContextPointer, this info was not inherited by popup menus generated in this layout. While operators from regular buttons work fine, the data is missing in operators from menus and such. This patch copies the bContextStore from buttons to the new uiLayout used for popups.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@45375 954f8c5b-7b00-dc11-b283-0030488c597c