Add zlib for MinGW32
not sure if the externals I use are safe enough, if a lib version changes in MSVC MinGW may break without notice. Still I guess since we do use a common dll for both it would be the same issue with the old scheme anyway. And if the need arises to separate, we can always do it later.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@45852 954f8c5b-7b00-dc11-b283-0030488c597c