Cleaned up glreport.py (pardon my dust)
See the README.txt for details of how to read the generated report.

Consolidated the 600+ extension files into a set of JSON style Python files.
Not much of a space reduction, but definitely a reduction in clutter.

People are free to edit the .gl files.  The warning in the files against editing is for people who are generating new .gl files from other sources.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@46340 954f8c5b-7b00-dc11-b283-0030488c597c