Python/context: python could get invalid bpy.data in scene update handler after undo.
The way this got updated from the context is a bit unreliable, and for handlers
the update couldn't happen because there is no context passed in. Now it's
updated from setup_app_data, which is where the change actually happens. I left
in the other updates to be sure but they should not be needed anymore.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@46437 954f8c5b-7b00-dc11-b283-0030488c597c