Reroute nodes, by Jeroen Bakker (patch #28443).
By holding shift and "cutting" a node link a new reroute helper node can be inserted. This consists of a single socket that can be used to insert additional connection points into a link. This can be used to keep a connection point in the tree when  deleting a node, or to control the path of long connections for layout cleanup.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@47319 954f8c5b-7b00-dc11-b283-0030488c597c