Fix [#31714] Pinning object context causes texture list to disappear
Problem was py code of main texture panel was not doing any check on the pinned id, assuming it managed the textures itself - but this is not the case of the Object datablock...

All work actually done by Sergey, was just missing the Lamp specific case. Checked both in code and with tests, quite sure all cases are now correctly handled!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@48151 954f8c5b-7b00-dc11-b283-0030488c597c