Possible fix for [#32141] Crash when using a mask as the factor input for a color combine (mix) node with render resolution at 100%
Seems to be that the MaskNode has been created as a complex node. But no
complex features were used. Converted the execute pixel to simple
execution. And it sees that the crash does not happen.

Not sure if it is the issue is solved. I am going to let the user retest
with this revision.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49069 954f8c5b-7b00-dc11-b283-0030488c597c