Fix #32118, Reroute nodes inside groups get created in wonky places.
Previous attempt was not taking into account parent nodes of the group nodes (i.e. frames). The nodeFromView function should be used to convert between relative node location to absolute view space. Also added this in the node_add_node helper function, which is used in a number of operators for adding nodes in a specific location.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49085 954f8c5b-7b00-dc11-b283-0030488c597c