Fix #32004: Up/down arrow keys can move index out of actual range of template_list
Fixed by clamping current index value to 0..list_size-1 range in list event
handling function. This shouldn't give any regressions since this clamping\
happens only after template was already displayed so items counter should be
correct here.

It is still possible to set index to meaningless value by directly changing
the index value via python, but that's not in the scope of interface engine
and not currently considered a bug.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49218 954f8c5b-7b00-dc11-b283-0030488c597c