Initial support for numpy in scons
Assumes numpy is installed to the BF_PYTHON/site-packages/numpy directory,
could be tweaked further, but this should be enough to setup release
building environment.

Seems blender can't import numpy, but that doesn't seem to be scons issue,
the same happens here with cmake too. Would ask Campbell to help looking
into this.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49303 954f8c5b-7b00-dc11-b283-0030488c597c