Tie compositor will now update render result when changing node setup
Issue was caused by the way how render result was acquiring -- pointer
to render data was used to find needed render descriptor. It's not
reliable since render contains copy of scene's render data, not pointer
to this data.

Use node scene's id name for render result acquiring, the same way
as it was done in old compositor system.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49450 954f8c5b-7b00-dc11-b283-0030488c597c