Sequencer: initial support of refreshing only changed sequences
Before this the the whole sequencer cache would be invalidated
when hanging sequence settings.

This was completely annoying because changing color balance
settings would re-load image file for which color balance
is happening on every change,

In fact it's still an issue if color balance is changing for
image strip itself, but if this strip has got effect and
color balance is changing for it file wouldn't be reloaded.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49702 954f8c5b-7b00-dc11-b283-0030488c597c