Sequencer: support for masked color balance
This implements option which could be used to color balance only
specified area. Currently done by adding Mask input to Adjustment
effect. Affects on color balance and multiply settings.

Supporting masked saturation control is in the list, not supported
in this commit.

Also show value slider in the right of color wheel.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@49806 954f8c5b-7b00-dc11-b283-0030488c597c