Fix #32201: particle size compatibility broken for object/group duplication.
After 2.63 there was a bugfix to take object scale into account for the duplicated
objects, but this breaks compatibility on earlier files. Now there is an option to
control if the scale should be used or not.

Scale is used by default on newer files, and not used on older ones.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@50116 954f8c5b-7b00-dc11-b283-0030488c597c