Fix #31539, painting in image editor while in object mode does not update mipmaps.
While we could disable/enable mipmaps on stroke begin/end, it is a bit hacky (but worthy of consideration for later) for my taste just to paint in the image editor. Instead we generate mipmaps on the fly. Since we can update texture levels below the first only with GPU mipmapping, partial update when painting in the image editor will actually work only with GPU mipmapping from now on (which is fast enough I hope not to get any lags!).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@50730 954f8c5b-7b00-dc11-b283-0030488c597c