Fix #32625 Zero area-sized faces caused jittered part. distrib. to crash
When appling a particle system to a face whose area size is zero,
the jitter distribution failed. fmod() produces a NaN value in this
case. This commit simply checks if the jitter offset (I guess that is
"jitoff" means) and only call psys_uv_to_w() if it's a real floating
point number.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@50831 954f8c5b-7b00-dc11-b283-0030488c597c