Fix #32091: Crop and offset coordinates changes proxy render settings on image strips
Made offset and crop working in final resolution space, meaning that visual result
would be the same when using different proxy resolution in sequencer preview and
different resolution percentage when rendering.

It could break some files designed to with render percentage not equal to 100%
using wrong image crop and offset values. But this would also make it easier to
setup scene and render it with different percentage (useful for preview renders
for example)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@50901 954f8c5b-7b00-dc11-b283-0030488c597c