Autokey warning - Only show for the active region
In response to some of the feedback, I've taken a second look at the situation
when multiple views are open, and indeed in that situation having it display in
every region and blinking was quite overwhelming (admittedly, I've mainly been
testing on single-view setups). Now it only shows for the region that was used
for initiating the transform.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51214 954f8c5b-7b00-dc11-b283-0030488c597c