Graph Editor: Added a filtering option for Drivers mode to only show F-Curves with errors
This filtering option is useful when rigging and you want to figure out if any
of your drivers are not functioning, and/or which one(s) are not, so that you
can go through fixing them. It saves you from having to check on each one
individually, or going into the console to try to infer which ones are not
working.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51246 954f8c5b-7b00-dc11-b283-0030488c597c