Bugfix [#32865] Usercounts for World Textures not decremented after preview render
When adjusting settings for world textures (with Both/World preview modes),
every tweak would result in the usercount of the texture increasing. As a
result, before long the texture would claim to have over 100 users. Fortunately,
this only appeared to be just a cosmetic issue (i.e. no real memory leak here),
though it was a bit unsettling.

NOTE: this is still a bit glitchy, as now we have flickering when updating
texture settings - the texture still temporarily has a second user during
preview rendering.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51288 954f8c5b-7b00-dc11-b283-0030488c597c