Cycles: object motion blur enabled, so in addition to camera motion, moving objects in the scene will also cause motion blur.
This change does come with a bit of a slow down to the CPU rendering kernel even
with motion blur disabled, due to extra overhead in handling of object matrices.
It's a few percentages on simpler scenes, not so noticeable on more complex ones.
With motion blur enabled rendering is of course also slower as would be expected,
though from testing especially GPU rendering handles it quite well.

This does not support motion blur from deforming objects yet, only translation,
scale and rotation. Deformation blur is probably for another release.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51363 954f8c5b-7b00-dc11-b283-0030488c597c