Fix #32705: Esc a value change doesn't recalc compositor
Seems the issue was caused by G.is_break set to Truth on both escape
press and release. This ended up in situation when after press event
compositor was tagged to redraw and new job was started. On escape
release this compositor job was cancelled.

Made it so G.is_break is setting on escape press event only.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51421 954f8c5b-7b00-dc11-b283-0030488c597c