Fix #32837: DDS images flipped in viewport compared to render. This happened after adding compressed DDS texture loading.
DDS images can be flipped compared to the Blender standard, however we do not
unflip them because we also don't flip compressed textures. If we would flip
those we'd need to uncompress, flip and recompress them, and so losing the
speed benefit that you get from using them. Users are expected to save DDS
image in OpenGL compatible format.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51546 954f8c5b-7b00-dc11-b283-0030488c597c