Adding a pixelate node.
This makes it possible to create pixelized scale in the Tile compositor.

Just append the node in front of a scale node or where you want the pixelization to take place.

There were some bugs on this subject, but they used the work around to add a blur size of 0 in the place where they need the pixelization.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@51572 954f8c5b-7b00-dc11-b283-0030488c597c