Correction to previous commit
pBVH nodes buffers could share some pointers from pBVH itself,
and after updating pointers in pBVH we need to recreate this buffers.

This starts looking a bit awkward and seems it's mainly because of
assumptions made when was working on sculpting on deformed mesh.
Would revision such a sculpting and would try to make it more clear
later.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@52008 954f8c5b-7b00-dc11-b283-0030488c597c