Install dependencies: disable OSL for now
It's really horror even for me to compile it on release environment,
i do not want anybody to spend time trying to support this lib in
automatic script or make it so user's are easily frustrated by some
hack added to OSL repository.

If you REALLY want to build OSL with this script, set BUILD_OSL to
truth (it's in the top of the script).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@52299 954f8c5b-7b00-dc11-b283-0030488c597c