Compositor should never add notifiers by himself, notifiers should be added from main thread using job update callback.
Added new execution-time callback to bNodeTree which marks job to be updated.

The code here could be a bit not so obvious because in some cases job update
callback need to merge local tree, but it's only needed for old compositor
system which is gonna to be removed soon, so decided not to bother with
cleanup now. Removing old compositor system will also allow to drop stats_draw
callback from bNodeTree.

This should fix following bugs:

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@52679 954f8c5b-7b00-dc11-b283-0030488c597c