Bug fix, irc submitted:
On saving a first .blend after startup, the file would load back as if it was not
saved (showing no name in header). 

The whole FILEFLAGS and G.relabase_valid and G.file_saved etc is messy.
This commit fixes issues, but only adds more mess :)

Will discuss a nicer implementation of all of this.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@53011 954f8c5b-7b00-dc11-b283-0030488c597c