Version patch for reading old files (2.50 and before).
In 2.50 a new convention was added to save the filename in the .blend itself. 
This to allow recovery of temp saves.

In current svn, it made old files open as if it was a saved home file (not
storing the name in header, or in file history).

Note: file handling for all recovery, remapping etc is in need for cleanup.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@53084 954f8c5b-7b00-dc11-b283-0030488c597c