Better error report when no camera is found.
While trying to render a RenderLayer from a linked scene, I couldnt tell why it wasnt doing anything. It just stopped rendering. Until I look on the console and see the No Camera print.

This commit replaces the simple terminal printf with an actual error report, that shows in both terminal and to the user:
http://pasteall.org/pic/show.php?id=42499

I believe this commit is pretty harmless, and only adds functionality. But feel free to remove it.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@53349 954f8c5b-7b00-dc11-b283-0030488c597c