New feature
Patch [#33445] - Experimental Cycles Hair Rendering (CPU only)

This patch allows hair data to be exported to cycles and introduces a new line segment primitive to render with.

The UI appears under the particle tab and there is a new hair info node available.

It is only available under the experimental feature set and for cpu rendering.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@53373 954f8c5b-7b00-dc11-b283-0030488c597c