Support normalized displacement maps in cases maximal distance is not set
This will calculate maximal distance automatically and normalize displacement
to it. Before this change normalization will not happen at all in cases max
distance is not set manually.

This affects on "regular" baker only, there are still some fixes to come for
multiresolution baker, but that could be solved separately.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@53947 954f8c5b-7b00-dc11-b283-0030488c597c