Fix for a nasty glicth found by Irie Shinsuke in Freestyle branch.
Headerless panels are not supposed to be closed ever. But if user saves a blend with a stardard panel closed, then dev decides to make this panel headerless, when user open again its blend, the panel is closed and has no more header, so it becomes invisible!

This commit simply checks, at draw time, that a headerless panel is never closed (and repoen it if necessary)!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@54141 954f8c5b-7b00-dc11-b283-0030488c597c