Move opencl and reigidbody from source/blender/ to intern/
This modules does not depend on any blender-specific data
structures or algorithms and due to our policy better be
placed to intern/

Shall be no functional changes, tested CMake and SCons on
Linux, hopefully other platforms will work as well.

P.S. SVN history shall be preserved for the files.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@54247 954f8c5b-7b00-dc11-b283-0030488c597c