Reports #34042 and #33749
Tweak for new option for Relative Bone parenting
(which transforms child object based on rest pose, so you can change bones 
in editmode to define pivot)

In the original commit it was made default, but that was too invisble for
users. Now it's an option in the Make Parent menu to choose. Communicates
a new feature better.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@54494 954f8c5b-7b00-dc11-b283-0030488c597c