Cycles GPU rendering: * Deprecate computing capability 1.3 (sm_13)
This commit disables auto build of sm_13 CUDA platform, which means that starting with Blender 2.67, we don't support sm_13 devices anymore. It has become difficult to support that and it was already feature incomplete (no render-passes, AO, Multi Closure etc).

It's still possible to manually enable sm_13 for own tests, but building might break in the future. 

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@54706 954f8c5b-7b00-dc11-b283-0030488c597c