rigidbody: Add motor constraint
It's implemented as a separate constraint instead of adding properties
to the existing constraints.
Motors only apply linear and angular impulses and don't limit the
movement of rigid bodies, so it's best to use them in conjunction with
other constraints to limit the degrees of freedom.

Thanks to Markus Kasten (markus111) for the initial patch.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@54798 954f8c5b-7b00-dc11-b283-0030488c597c