Fix #34322: cycles crash with (undo) save during threaded render. The mesh save code was modifying pointers in the Mesh which gave crashes with another thread accessing the data at the same time. This could crash other threaded operations like blender internal render or physics baking too but was less likely.
As a solution I've now changed the save code that it does not modify the mesh
data structure in place but rather a copy, as undo file saving should probably
be fully read-only regardless of how an improved threading architecture might
work.

Thanks to Sergey for tracking down the cause of this crash.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@55991 954f8c5b-7b00-dc11-b283-0030488c597c