Bug fix #34896
The feature "Keep Session" was also loading that session when you double-click
on a .blend to open it, or when a .blend file was on commandline.

Moved this feature to the main() in creator.c, so it can check on it properly, skipping the
kept session when a file was loaded.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56006 954f8c5b-7b00-dc11-b283-0030488c597c