Fix for access freed scene when doing freestyle render
It's more like an ownership thing, but it's not so much
clear who actually is an owner of temporary scene, so
for now just ensure Render does not have link to it
after strokes are rendered.

There're also suspicious thing related on why separate
RE_free_database is needed for freestyle renders, but
that's also another story.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56136 954f8c5b-7b00-dc11-b283-0030488c597c