Bugfix #34895
Error in shadowbuffer with halo effect. It had render error when you look straight into
the lamp, with something in front of lamp.

The code that was sampling the shadowbuffer exited too soon (for ray going perpendicular to buffer it should take at least 1 sample).

Code from 90ies. And still got a bug fix. Whoa :)



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56145 954f8c5b-7b00-dc11-b283-0030488c597c