Fix #35051: when viewing a complex node setup in the material properties, the properties were continuously redrawing, which slowed down everything else.
The problem was integer overflow, with a short only capable of storing values
up to 32767. Note that sockets are collapsed by default since the previous
release, and that's it's not very useful to edit such complex node setups in
the properties editor, it's mainly meant for simple setups or group nodes to
present just a few sockets.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56210 954f8c5b-7b00-dc11-b283-0030488c597c