game engine: Use bullet's default contact processing theshold
Blender's old default of 1 was too small here.
Now we don't override the default.
If it's really needed to use a smaller value here, the property should
be first exposed in the ui, but don't think this is necessary.

Fixes [#35076] Dynamic objects have periodic glitch in velocity when on an incline

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56302 954f8c5b-7b00-dc11-b283-0030488c597c