Fix #35149: solidify modifier + vertex parent not working after going in and out of editmode on the child object.
Problem was that the object custom data mask was not taken into account when
rebuilding the derivedmesh in some cases, which is needed for the derivedmesh
to contain the mapping back to the original vertices. Now this data mask is
used for any derivedmesh build that will be cached.

Also problematic was that the datamask for the active object was applied to
all objects in the scene, which caused the parent object to be recalculated
when it didn't need to be. Now this datamask is only used for the active object.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56447 954f8c5b-7b00-dc11-b283-0030488c597c