multiview: Compositor support
Right now the view loop happens in pipeline.
That means only explictly calling render will composite multiviews.

Implementation detail:
----------------------
I focused on the multilayer exr output node
(COM_OutputFileOperation.cpp).

And it is working fine (though always saving MultiPart exr,
and having one whole view per part).

Since the whole composite context is thrashed between view renders,
I'm storing the exr handle in the static ListBase in openexr_api.cpp:

When the first view runs, it creates a new handle.
When the second (or more) view runs, it loads the existent.
When the last view runs, it closes the handle.

This is necessary because we can't write to a MultiPart file in "sessions"
(as in, open the file and just overwrite one view).

Things not tackled yet:
----------------------
* FSA
* Viewer Node
* Image Node (right now it shows each view as a socket)
* SingleLayer exr file
* Non-exr files (more a design/UI thing)

* Global option to composite/render only one view
* Image Node: to map image views to scene views
* View Filter Node: to run a nodebranch when view =="left" ...

Note:
-----
I'm storing view as integer, but for Image Node and View Filter
Node I think I will need to store the actview as string.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch multiview
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   blenkernel/BKE_node.h
#	modified:   compositor/COM_compositor.h
#	modified:   compositor/intern/COM_CompositorContext.h
#	modified:   compositor/intern/COM_ExecutionSystem.cpp
#	modified:   compositor/intern/COM_ExecutionSystem.h
#	modified:   compositor/intern/COM_compositor.cpp
#	modified:   compositor/nodes/COM_ImageNode.cpp
#	modified:   compositor/nodes/COM_OutputFileNode.cpp
#	modified:   compositor/nodes/COM_RenderLayersNode.cpp
#	modified:   compositor/operations/COM_OutputFileOperation.cpp
#	modified:   compositor/operations/COM_OutputFileOperation.h
#	modified:   compositor/operations/COM_RenderLayersBaseProg.cpp
#	modified:   compositor/operations/COM_RenderLayersBaseProg.h
#	modified:   editors/space_node/node_edit.c
#	modified:   imbuf/intern/openexr/openexr_api.cpp
#	modified:   imbuf/intern/openexr/openexr_multi.h
#	modified:   imbuf/intern/openexr/openexr_stub.cpp
#	modified:   nodes/composite/node_composite_tree.c
#	modified:   nodes/composite/nodes/node_composite_outputFile.c
#	modified:   render/extern/include/RE_pipeline.h
#	modified:   render/intern/source/pipeline.c
#	modified:   render/intern/source/render_result.c
#	modified:   render/intern/source/rendercore.c
#	modified:   render/intern/source/zbuf.c
#