multiview: fix for passes not changing
basically the iuser->pass was set in image_buttons.c
but that wasn't the real pass. I created now a tmp_pass
to account for that.

The real iuser->pass is then set in the image.c
I could handle that in image_multi_cb() instead:

iuser->pass = iuser->pass_tmp + (iuser->view * BLI_countlist(&rr->views));

The nice thing of handling in image.c is that when we do 3d support I
think there will be the place to implement. We'll see.