Fix #35265: on OS X, pressing system shortcuts such as cmd+M or cmd+` would both insert text in the text editor and do the associated operation like minimizing the window or switching windows.
The code was always doing both without trying to ensure only one is done. Now we
integrate a bit better with the event handling and pass the event to NSApp, which
then decides to handle the event itself or pass it on to the window, from where
we then send it back to be handled.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56683 954f8c5b-7b00-dc11-b283-0030488c597c