Fix ctrl+(shift)+tab shortcut key not working on OS X after recent bugfix to interact better with system shortcuts.
This is a special shortcut for switching between views and does not get
delivered directly to our view when we pass it through the application key
event handling path. We only have a single OpenGL view, so there's no need to
pass it on to the application, instead just interpret it directly.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56728 954f8c5b-7b00-dc11-b283-0030488c597c