Fix #35443: Animation player doesn't play some movie formats
Need to initialize FFmpeg for animation player.

This is a bit confusing, but the only way to do this
properly and clear is to call IMB_ffmpeg_init from
playback_mode command line callback.

This is because this callback is blocking and
"usual" FFmpeg initializetion (which happens after
command line parsing) was never called.

Could be switched to separate FFmpeg initializtion
and debug level set but let's keep simple for now.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@56943 954f8c5b-7b00-dc11-b283-0030488c597c