multiview: save image from Image Editor working TODO: save compositor image
Also added a new flag to exr types: "Multiview" should be used to save
the multiview in one file, while "Multipart" is simply an option to
benefit from the EXR 2.0 format.

Food for thought:
==================
Maybe the "Multiview" option should be visible for all image types.
Otherwise there is no way to save only the current view.
On the other hand, there is no way to save only "R" from an image, so it
may not be a problem.