Bugfix [#35382] NLA "Multiply" Blend Mode calculated incorrectly
The "Multiply" blending mode for NLA strips worked incorrectly. Instead of
modulating the influence of the current strip, it was in fact scaling the result
of the entire stack (with the strip applied). This caused problems when
influence = 0, as it was in fact muting everything instead of just controlling
the strip we are interested in.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@57345 954f8c5b-7b00-dc11-b283-0030488c597c