Fix for  * [#35724] Backdrop zoom can be set to a very small value, making the backdrop disapear.
There were checks in the drawnode that needed to be placed in the readfile.
The checks checked if the zoomlevel was 0.0, then it was defaulted to 1.0, but the zoomvalue had a minimum limit of 0.01, hence it did not work.
Moved the check to the readfile and checked for all values smaller then 0.02. These values are then reset to 1.0

Jeroen & Monique
 - At Mind -


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@57453 954f8c5b-7b00-dc11-b283-0030488c597c