multiview: refactor or drawdata (wm_draw.c) as per Brecht suggestion all drawdata could be stored in one single drawdata.
I went beyond and implemented the drawdata as a Listbase. This way it
becomes easier to extend the draw code for more fancy display modes in
the future (fulldome ...)

One thing I didn't do, is to add the stereo_data and stereo_all inside
wmDrawData. So right now they are stored in sequence in the listbase.
I could move them to the struct, it may make the code more easy to read.
I will ask Brecht