Fix for Bug #35970: Freestyle + Particles = Crash.
The crash was caused by a reference of freed memory in add_halo_flare() that relied
on the global variable R to retrieve a list of objects.  The value of this variable
is invalidated by a nested execution of the Blender Internal renderer for Freestyle
stroke rendering, leading to the documented memory reference issue.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@58063 954f8c5b-7b00-dc11-b283-0030488c597c