Fix #36064: cycles direct/indirect light passes with materials that have zero RGB color components gave non-grey results when you might no expect it.
What happens is that some of the color channels are zero in the direct light
pass because their channel is zero in the color pass. The direct light pass is
defined as lighting divided by the color pass, and we can't divide by zero. We
do a division after all samples are added together to ensure that multiplication
in the compositor gives the exact combined pass even with antialiasing, DoF, ..

Found a simple tweak here, instead of setting such channels to zero it will set
it to the average of other non-zero color channels, which makes the results look
like the expected grey.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@58086 954f8c5b-7b00-dc11-b283-0030488c597c