Expose node.select exec callback to an operator
Internally it was only invoke callback set for an
operator template. This invoke was setting such
properties as mouse_x and mouse_y and was calling
an exec function.

This meant that t seemed to be really easy to
use node.select operator from by giving a mouse
positions, but in fact it wasn't possible (because
it requires exec callback)

This commit sets operator's template exec callback,
which makes it possible using node.select from
python.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@58191 954f8c5b-7b00-dc11-b283-0030488c597c