BGE: Flipping vsync constants so VSYNC_ON is 0.
This will make transitions from older versions of Blender easier since VSYNC_ON
will be the default. This could have been changed in a do_version, but the vsync
code has yet to see an official release, so I figured this would be a bit nicer.
Also, this makes VSYNC_ON the default for new scenes as well.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@59194 954f8c5b-7b00-dc11-b283-0030488c597c