Attempt fix for #36688.
Curves may not be not initialized when called from python. C code
explicilty says that curvemapping_initialize should be called prior to
evaluating the curve, however the curve clip rectangle is not available
when calling evaluation on the curvemap. This is not possible unless we
force the evaluation on CurveMapping level, not on CurveMap level.

For now just pass a rectangle with the x boundary values of the curvemap
for evaluation to avoid the crash.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@59965 954f8c5b-7b00-dc11-b283-0030488c597c