bugfix [#36757] Vertex coloring in preview window meshes default to black, makes preview useless
Adding 'Col' vertex color layer to preview cycles meshes

The artist could simply disconnect the VCol node from the nodetree for
the preview, but  it should be harmless to add this to the preview meshes.

As long as the user stick to the default vertex color name (and is happy
with seing a all-white channel) it should be ok.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@60206 954f8c5b-7b00-dc11-b283-0030488c597c