Fix duplicated key in ghash assert caused by object orco
Code was rather confusing, get_object_orco was only
checking for orco in cache for some objects and was
actually allocating orco for other objects.

Now made it so get_object_orco always only checks
the cache and only call set_object_orco if there's
no orco for it yet.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@60772 954f8c5b-7b00-dc11-b283-0030488c597c