Bugfix [#36844] Cannot set Restrict Frame Start for FModifiers until Frame End has been adjusted
Previously, the RNA settings tried to strictly enforce the constraint that the
start frame must be less than the end frame. However, this behaviour was
problematic, as it meant that you had to firstly move the end frame to its new
(higher) value, before moving the start frame. The same also applied in the
opposite direction.

Now, this behaves in the same way that the scene start/end buttons work: if the
new start frame is past the end frame, the end frame is "pushed" along to be the
same value as the start frame. The same applies in the opposite direction.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk@61105 954f8c5b-7b00-dc11-b283-0030488c597c